==> R - hei_aim3a/1.2_site_type.Rmd <==
---
title: "Selection of sites by type"
author: "Magali Blanco"
date: ' `r Sys.time()` '
output:
  html_document:
    df_print: paged
    toc: yes
    toc_depth: '3'
    number_sections: true
    toc_float: true
    collapsed: false
    smooth_scroll: false
editor_options: 
  chunk_output_type: console
---


```{r, echo=F, warning=F}
##################################################################################################
# SETUP
##################################################################################################
knitr::opts_chunk$set(echo = F, 
                      cache=F, cache.comments = F, 
                      message = F, warning = F#, tidy.opts=list(width.cutoff=60), tidy=TRUE 
                      )  

# Clear workspace of all objects and unload all extra (non-base) packages
rm(list = ls(all = TRUE))
if (!is.null(sessionInfo()$otherPkgs)) {
  res <- suppressWarnings(
    lapply(paste('package:', names(sessionInfo()$otherPkgs), sep=""),
           detach, character.only=TRUE, unload=TRUE, force=TRUE))
}

pacman::p_load(tidyverse)    

source("functions.R")
set.seed(1)

latest_version <- "v3_20230321" 
dt_path <- file.path("Output", latest_version)

if(!dir.exists(file.path(dt_path, "campaign visit samples"))) {dir.create(file.path(dt_path, "campaign visit samples"))}

image_path <- file.path("..", "Manuscript", "Images", "v4", "SI")

# ggplot settings
theme_set(theme_bw())
theme_update(legend.position = "bottom")

##################################################################################################
# LOAD DATA
##################################################################################################
# all data 
keep_vars <- readRDS(file.path(dt_path, "keep_vars.rda"))
        
stops_w <- readRDS(file.path(dt_path, "stops_used.rda"))
# all-data estimates
all_data_annual <- stops_w %>%
  group_by(location) %>%
  summarize_at(all_of(keep_vars), ~mean(., na.rm=T)) %>%
  mutate(design = "full",
         version = "all data", #"all training data",
         campaign=1) %>%
  ungroup()

# site type categorization (from Amanda)
pls_categories <- read.csv(file.path("data", "pls_cats.csv"))[-1] 
high_sites <- filter(pls_categories, pls_cats =="high") %>% pull(location)
med_sites <- filter(pls_categories, pls_cats =="med") %>% pull(location)
low_sites <- filter(pls_categories, pls_cats =="low") %>% pull(location)

##################################################################################################
# COMMON VARIRABLES
##################################################################################################
# number of simulations
sim_n <- 30

visit_count <- seq(2,22, 2)
site_type_levels <- paste("H", rev(visit_count), "L", visit_count)
saveRDS(site_type_levels, file.path(dt_path, "site_type_levels.rda"))

default_count <- 12

keep_vars <- readRDS(file.path(dt_path, "keep_vars.rda"))

```

# summarize site types


```{r}
print("site variability categorizations")
table(pls_categories$pls_cats) # 90*2/309 = 58% of sites changing 

pls_categories %>%
  mutate(
    pls_cats = ifelse(pls_cats=="med", "medium", pls_cats),
    pls_cats = str_to_title(pls_cats),
    pls_cats = factor(pls_cats, levels=c("Low", "Medium", "High"))
    ) %>%
  pivot_longer(cols = c(iqr, predicted_iqr)) %>%
  mutate(
    name = ifelse(name == "iqr", "Observed IQR",
                  ifelse(name == "predicted_iqr", "From PLS, Used in Sampling Designs", NA))) %>%
  
  group_by(pls_cats, name) %>%
  summary_table(var = "value") %>%
  ggplot(aes(x=pls_cats, fill=name)) +
  geom_boxplot(stat="identity", aes(ymin=Q10, lower=Q25, middle=Q50, upper=Q75, ymax=Q90)) +
  
  # ggplot(aes(x=pls_cats, y=value, fill=name)) + 
  # geom_boxplot() + 
  labs(y="PNC IQR (pt/cm3)", x="Predicted Site Variability (IQR)",
       fill = NULL
       )

ggsave(file.path(image_path, "site_type_iqr.png"), width = 6, height = 4)

```



# Sample by site type

* refernce is 12 visits/site x 309 sites = 3708 total stops.  
* go as far down/up as 4/20 (or 2/22) visits per site

```{r}
site_type_visits_df <- data.frame()

set.seed(1)
# v = 1
for(v in 1:length(visit_count)) {
  
  low_var_count <- visit_count[v]
  high_var_count <- rev(visit_count)[v]
  
  # c=1
  temp <- lapply(1:sim_n, function(c){
    
    low_var_sites <- stops_w %>%
      filter(location %in% low_sites) %>% 
      group_by(location) %>%
      slice_sample(n = low_var_count, replace=T) 
    
    med_var_sites <- stops_w %>%
      filter(location %in% med_sites) %>% 
      group_by(location) %>%
      slice_sample(n = default_count, replace=T) 
      
    
    high_var_sites <- stops_w %>%
      filter(location %in% high_sites) %>% 
      group_by(location) %>%
      slice_sample(n = high_var_count, replace=T)
    
    rbind(med_var_sites, low_var_sites)%>%
      rbind(high_var_sites) %>%
      mutate(campaign = c,
             design = "site type",
             version = paste("H", high_var_count, #"med", default_count, 
                             "L", low_var_count)) %>%
      ungroup()
      
    }) %>%
    bind_rows()
  
  site_type_visits_df <- rbind(site_type_visits_df, temp)
  
  }
# save visit samples
saveRDS(site_type_visits_df, file.path(dt_path, "campaign visit samples", "site_type.rda") )

# #calculate annual averages
site_type_visits_annual <- site_type_visits_df %>%
  select(design, version, campaign, location, keep_vars) %>%  
  pivot_longer(cols = keep_vars, names_to = "variable") %>%
  group_by(design, version, campaign, location, variable) %>%
  summarize(visits = n(),
            mean = mean(value, na.rm = T)) %>%  
  pivot_wider(names_from = "variable", values_from = "mean") %>%
  ungroup()

# save data 
saveRDS(site_type_visits_annual, file.path(dt_path, "annual_site_type.rda") )

```

# Summarize averages 

### --> scatterplot; col by type?

```{r}
version_levels <- c(unique(all_data_annual$version), 
                    paste("H", rev(visit_count),
                          "L", visit_count))

compare <- bind_rows(site_type_visits_annual, all_data_annual) %>%
  mutate(version = factor(version, levels = version_levels)) 

compare %>%
  ggplot(aes(x=version, y=ns_total_conc)) + 
  geom_boxplot()


# compare


```


# compare annual averages 

### --> see email - what descriptives to show?

```{r}
compare_w <- site_type_visits_annual %>%
  left_join(select(all_data_annual, location, all_data_ns_total_conc =ns_total_conc)) %>% 
  left_join(select(pls_categories, location, pls_cats)) %>%
  mutate(version = factor(version, levels = version_levels),
         pls_cats = factor(pls_cats, levels=c("low", "med", "high"))) 

compare_w %>%  
  ggplot(aes(x=all_data_ns_total_conc, y=ns_total_conc, col=pls_cats)) + 
  facet_wrap(~version) + 
  geom_abline(intercept = 0, slope = c(1, 0.7, 1.3), linetype=2, alpha=0.5) +
  geom_point(alpha=0.2) + 
  geom_smooth() +
  labs(title = "comparison of annual average site estimates")
  
print("each boxplot has 309 sites x 30 campaigns (9,270) data points")
compare_w %>%
  mutate(difference = ns_total_conc - all_data_ns_total_conc) %>%
  ggplot(aes(x=pls_cats, y=difference, col=pls_cats)) + 
  facet_wrap(~version) + 
  geom_hline(yintercept = c(0, -2e3, 2e3), linetype=2, alpha=0.5) +
  geom_boxplot() + 
  labs(y= "difference in annual site concentration (pt/cm3): all data vs campaign",
       title = "distribution of site estimate differences by design version",
       subtitle = "dashed lines are at 0, +-2k")
  
```

 
# campaign performances 

```{r}
 
```
