==> R - hei_aim3a/9_comparisons.Rmd <==
---
title: "9_comparisons"
author: "Magali Blanco"
date: "2024-05-24"
output: html_document
editor_options: 
  chunk_output_type: console
---

```{r setup}
knitr::opts_chunk$set(echo = T, 
                      cache=F, cache.comments = F, 
                      message = F, warning = F, 
                      tidy.opts=list(width.cutoff=60), tidy=TRUE#, fig.height = 10, fig.width = 10
                      )  

# Clear workspace of all objects and unload all extra (non-base) packages
rm(list = ls(all = TRUE))
if (!is.null(sessionInfo()$otherPkgs)) {
  res <- suppressWarnings(
    lapply(paste('package:', names(sessionInfo()$otherPkgs), sep=""),
           detach, character.only=TRUE, unload=TRUE, force=TRUE))}

pacman::p_load(tidyverse, 
               knitr, kableExtra#, ggpubr, #ggarrange()
               )    

dt_path <- file.path("Output", readRDS(file.path("Output", "latest_dt_version.rda")))
dt_path <- file.path(dt_path, "epi","20240725")

set.seed(1)

source("functions.R")
out_path <- file.path("..", "Manuscript", "Images", "v5", "HEI Report")

##################################################################################################
# LOAD DATA
##################################################################################################
#all model coefficients - want ap & age effects
coef_all <- readRDS(file.path(dt_path, "model_coefs_all.rda")) %>%
  filter((grepl("nstot_all|pncnoscreen_all", model)))

# ## stationary and LCM
# st_ref_model <- "full_ST" #ST 
# sp_ref_model <- "full_SP" # SP 
# # ## no LCM
# # st_ref_model2 <- "fptv_no" #ST 
# # sp_ref_model2 <- "fp_no" # SP 


coef_lcm_all <- readRDS(file.path(dt_path, "models_lcm_all.rda")) %>%
  filter(grepl("full_", model) |
           
        
           model %in% c("no22_1", "no2_SP_2_1", #gold w/o historical - all monitors, ST & SP models
                        "no26_1", "no2_SP_6_1" #no homes - AQS + 3 snapshot
                        ))


coefs <- rbind(coef_all, coef_lcm_all) %>%
  filter(grepl("avg_0_5_yr|age", term)) %>%
  mutate(term = ifelse(term == "avg_0_5_yr", "ap",
                       ifelse(term == "visit_age_centered75", "age", term)),
         label = paste0(round(est, 3), " (", round(lower, 3), ", ", round(upper, 3), ")")
         )  %>%
   select(term, est, label, est_lower=lower, est_upper=upper,
         model) %>% 
  pivot_wider(names_from = term, values_from = c(est, label,
                                                 est_lower, est_upper,
                                                 )) %>% 
  
  select(model, contains("ap"), contains("age"), -contains("label"), contains("label"), everything()) %>% 
  mutate(
    ap_age_equivalent = est_ap/est_age,
    ap_age_equivalent_lower = est_lower_ap/est_age,
    ap_age_equivalent_upper = est_upper_ap/est_age,
    
    ap_month_equivalent = ap_age_equivalent*12,
    ap_month_equivalent_lower = ap_age_equivalent_lower*12,
    ap_month_equivalent_upper = ap_age_equivalent_upper*12,
  )

write.csv(coefs, 
          file.path(out_path, "hei_table3.2_age_equivalents.csv"), 
          row.names = F)

```

# compare all-data health ests
compare beta & age coefficients for all-data estimates
```{r}
coefs %>%
  select(model, label_ap, label_age, ap_month_equivalent,
         ap_month_equivalent_lower, ap_month_equivalent_upper
         ) %>%
  kable(caption = "Estimated associations between air pollutants (PNC [1,900 pt/cm3], PM2.5 [1 ug/m3]) and one year of aging with baseline CASI-IRT score, as determined by reference models (PNC all-data stationary mobile monitoring model, PM2.5 models with LCMs). Air pollution associations are scaled relative to the age association to contextualize their effects.",
        digits=3, 
        col.names = c("Model", "Air Pollution", "Age", "Air Pollution Month Equivalent", "..Month Equiv Lower", "..Month Equiv Upper")
        ) %>%
  kable_styling()


```
