==> R - hei_aim3a/9_ml_summary.Rmd <==
---
title: "Epi Results using Machine Learning UFP Exposure Models"
author: "Magali Blanco"
date: ' `r Sys.time()` '
output:
  html_document:
    df_print: paged
    toc: yes
    toc_depth: '3'
    number_sections: true
    toc_float: true
    collapsed: false
    smooth_scroll: false
editor_options: 
  chunk_output_type: console
---

# Setup 

```{r setup, include=FALSE}
knitr::opts_chunk$set(echo = F, 
                      cache=F, cache.comments = F, 
                      message = F, warning = F, 
                      tidy.opts=list(width.cutoff=60), tidy=TRUE#, fig.height = 10, fig.width = 10
                      )  

# Clear workspace of all objects and unload all extra (non-base) packages
rm(list = ls(all = TRUE))
if (!is.null(sessionInfo()$otherPkgs)) {
  res <- suppressWarnings(
    lapply(paste('package:', names(sessionInfo()$otherPkgs), sep=""),
           detach, character.only=TRUE, unload=TRUE, force=TRUE))}

pacman::p_load(tidyverse, kableExtra)    

dt_path <- file.path("Output", readRDS(file.path("Output", "latest_dt_version.rda")))

set.seed(1)

source("functions.R")
output_data_path <- file.path(dt_path, "epi", "20240725")

image_path <- file.path("..", "Manuscript", "Images", "v4", "other", "ml")
if(!dir.exists(image_path)) {dir.create(image_path, recursive = T)}
# ggplot settings
theme_set(theme_bw())
theme_update(legend.position = "bottom")

##################################################################################################
# LOAD DATA
##################################################################################################
model_levels <- c("UK-PLS", "RF", "TPRS", "RF-TPRS", "TPRS-RF", "SpatRF-PL", "SpatRF-NP")
comparison_models <- setdiff(model_levels, "UK-PLS")

# person-level exposures 
cs_ml <- readRDS(file.path(output_data_path, "dt_for_cross_sectional_analysis_machine_learning.rda")) %>%
  rename(exposure = avg_0_5_yr) %>%  
  select(-starts_with("avg_")) %>%
   mutate(model = gsub("^u", "", model),
         model=str_to_upper(model),
         # model = ifelse(grepl("pls", model, ignore.case=T), "UK-PLS", model),
         # model = factor(model) %>% relevel(ref="UK-PLS")
          model = ifelse(grepl("pls", model, ignore.case=T), "UK-PLS",
                        ifelse(grepl("RT", model, ignore.case=T), "RF-TPRS",
                               ifelse(grepl("TR", model, ignore.case=T), "TPRS-RF",
                                      ifelse(grepl("SPATPL", model, ignore.case=T), "SpatRF-PL", 
                                             ifelse(grepl("SPATCV", model, ignore.case=T), "SpatRF-NP", model))))),
         model = factor(model, levels = model_levels)
         )
  
# health effect estimates & CW
model_coefs_ml <- readRDS(file.path(output_data_path, "model_coefs_ml.rda")) %>%
  mutate(variable=="pnc_noscreen",
         ref_model = ifelse(model=="upls", "Reference Model", "Alternative Machine Learning Model"),
         ref_model = factor(ref_model, levels=c("Reference Model", "Alternative Machine Learning Model"))
         ) %>%
  label_pollutants() %>%
  mutate(model = gsub("^u", "", model),
         model=str_to_upper(model),
         model = ifelse(grepl("pls", model, ignore.case=T), "UK-PLS",
                        ifelse(grepl("RT", model, ignore.case=T), "RF-TPRS",
                               ifelse(grepl("TR", model, ignore.case=T), "TPRS-RF",
                                      ifelse(grepl("SPATPL", model, ignore.case=T), "SpatRF-PL", 
                                             ifelse(grepl("SPATCV", model, ignore.case=T), "SpatRF-NP", model))))),
         model = factor(model, levels = model_levels)
         )

```

# Model definitions

Si used stop readings from the unscreened P-TRAK

```{r}
model_coefs_ml %>%
  select(model, method, variable, instrument) %>%
  kable() %>% 
  kable_styling()

```

# Model R2
This should be in Si's 2023 paper (SI?) along with model descriptiosn? 

# Participant Exposure Predictions 

comparison of predicted 5 yr exposures at baseline

```{r}
# plot
cs_ml %>%
  pivot_wider(names_from = model, values_from = exposure) %>% 
  pivot_longer(cols = all_of(comparison_models)  
               ) %>%
  mutate(name = factor(name, levels=model_levels)) %>% 
  
  ggplot(aes(x=get('UK-PLS'), y=value, col=name)) +
  geom_point(alpha=0.05) +
  geom_smooth() +
  geom_abline(slope = 1, intercept = 0, linetype=2) + 
  labs(col="Machine Learning Model", y="Machine Learning Prediction (UFP pt/cm3)", x= "UK-PLS Prediction (UFP pt/cm3)")

ggsave(file.path(image_path, "5yr_exposure_comparisons.png"), width = 6, height = 6)

#table
cs_ml %>%
  group_by(model) %>%
  alt_boxplot(var="exposure") %>%
  rename(Q2.5=Qmin, Q97.5=Qmax) %>%
  kable(caption = "Distribution of UFP (pt/cm3) predictions", 
        digits = 0, format.args = list(big.mark=",")) %>% 
  kable_styling()

```

**Correlations**: exposure predictions from all models generate highly correlated predictions

```{r}

cs_ml %>%
  pivot_wider(names_from = model, values_from = exposure) %>%
  rename(ukpls = 'UK-PLS') %>%
  pivot_longer(cols = all_of(comparison_models)) %>%
  group_by(name) %>%
  summarize(correlation = cor(ukpls, value)) %>%
  kable(caption = "Correlation (R) of UK-PLS predictions with predictions from machine learning methods", 
        digits = 2,
        col.names = c("ML Exposure Model", "Correlation with UK-PLS")
          ) %>%
  kable_styling()

```

# Comparison of health effect estimates (CASI-IRT)

the resulting health estimates from using UK-PLS and various ML UFP exposure models is similar, with random forests & spatcv being the most different (~25-35% lower point estimate)

```{r}
ref_beta <- model_coefs_ml$est[model_coefs_ml$model=="UK-PLS"] #%>% round(2)

print("adjustd association between P-TRAK UFP (1k pt/cm3) and CASI-IRT")
model_coefs_ml %>%
  ggplot(aes(x=model, col=ref_model)) +  
  geom_hline(yintercept = ref_beta, linetype=1, alpha=0.5) +
  geom_hline(yintercept = 0, linetype=2, alpha=0.2) +
  geom_pointrange(aes(y=est, ymin=lower, ymax=upper), size=0.5, alpha=0.8) +
  
  scale_y_continuous(sec.axis = sec_axis(
    trans=~(.-ref_beta)/-ref_beta, name="% Difference from UK-PLS", labels = scales::percent)) +
  labs(col = NULL,
           x="Model",
           y = "Coefficient Estimate (95% CI)"
           )

ggsave(file.path(image_path, "ufp_health_estimate.png"), width = 6, height = 4)

```
