==> R - hei_aim3a/9_road_summary.Rmd <==
---
title: "Epi Results using Onroad Exposure Models"
author: "Magali Blanco"
date: ' `r Sys.time()` '
output:
  html_document:
    df_print: paged
    toc: yes
    toc_depth: '3'
    number_sections: true
    toc_float: true
    collapsed: false
    smooth_scroll: false
editor_options: 
  chunk_output_type: console
---

# Setup 

```{r setup, include=FALSE}
# Rscript -e 'rmarkdown::render("9_road_summary.Rmd", "html_document")'
knitr::opts_chunk$set(echo = F, 
                      cache=F, cache.comments = F, 
                      message = F, warning = F, 
                      tidy.opts=list(width.cutoff=60), tidy=TRUE)  

# Clear workspace of all objects and unload all extra (non-base) packages
rm(list = ls(all = TRUE))
if (!is.null(sessionInfo()$otherPkgs)) {
  res <- suppressWarnings(
    lapply(paste('package:', names(sessionInfo()$otherPkgs), sep=""),
           detach, character.only=TRUE, unload=TRUE, force=TRUE))}

pacman::p_load(tidyverse, kableExtra, ggpubr, sf, ggspatial,# ggpubr::ggarange()
               #ggmap
               osmdata, 
               rosm, ## For OpenStreetMap tiles (background)
               ggspatial, # annotation_spatial
               tigris,
               ggmap,
               colorblindr 
               #ggsci #scale_fill_jco()
               )    

config <- yaml::read_yaml("config.yaml")

dt_path <- file.path("Output", readRDS(file.path("Output", "latest_dt_version.rda")))

set.seed(1)

source("functions.R")
output_data_path <- file.path(dt_path, "epi", "20240725")

output_final <- "Output/final/onroad"

image_path <- file.path("..", "..", "Manuscript - Onroad data", "Images", "v5")

if(!dir.exists(image_path)) {dir.create(image_path, recursive = T)}
if(!dir.exists(file.path(image_path, "SI"))) {dir.create(file.path(image_path, "SI"), recursive = T)}
if(!dir.exists(file.path(image_path, "HEI Report", "Appendix"))){dir.create(file.path(image_path, "HEI Report", "Appendix"), recursive = T)}
   
# ggplot settings
theme_set(theme_bw())
theme_update(legend.position = "bottom")
 
##################################################################################################
# DEFINITIONS
##################################################################################################
design_levels <- c(
  "balanced routes",
  "balanced segments",
  "unbalanced segments",
  "unbalanced random clusters",
  "unbalanced sensible clusters",
  "unbalanced unsensible clusters",
  "unbalanced road type") %>% str_to_title()

#design_levels <- c("Route", "Road Type", "Unsensible Clusters", "Random Clusters", "Sensible Clusters", "Random Segments", "Balanced")

##################################################################################################
# FUNCTIONS
##################################################################################################

##################################################################################################
summary_table <- function(df, var){
  df <- df %>%
    rename(var = all_of(var)) %>%
    
    #calculate quantiles
    summarize(
      N = n(),
      Min = min(var),
      Q05 = quantile(var, 0.05),
      Q10 = quantile(var, 0.10),
      Q25 = quantile(var, 0.25),
      Q50 = quantile(var, 0.50),
      Q75 = quantile(var, 0.75),
      Q90 = quantile(var, 0.90),
      Q95 = quantile(var, 0.95),
      Max = max(var)) %>%
    ungroup()
  
  names(df)[names(df)==var] <- var
  
  return(df) 
  
}

##################################################################################################
# return results relevant for HEI 
# dt=all_predictions
keep_hei_results <- function(dt, keep_visits = c(main_visits, alt_visits)) {
 dt %>%
  filter(#grepl(paste(keep_visits, collapse = "|"), visits), 
         visits %in% keep_visits,
         design %in% hei_designs,
         !grepl("Random|Adj 2|Underwrite", version),
         cluster_type %in% c(NA, "cluster1")
         ) 
}
 
# return results relevant for main analyses
keep_main_results <- function(dt,
                              designs_to_keep = main_designs,
                              versions_to_keep = main_versions,
                              visits_to_keep=c(main_visits, alt_visits),
                              cluster_to_keep = main_cluster,
                              times_selections_to_keep = unique(model_eval2$time_selection)
                             ) {
  dt %>%
    filter(visits %in% visits_to_keep,
           design %in% designs_to_keep,
           version %in% versions_to_keep,
           cluster_type %in% c(NA, cluster_to_keep),
           time_selection %in% times_selections_to_keep
           ) 
}

##################################################################################################

# dt=model_eval
# dt=cw
add_labels <- function(dt){
  
  adjustments <- c("No Adjustment", 
                   "Fixed Site Temporal Adj", "Fixed Site Temporal Adj Random", 
                   "Underwrite Temporal Adj", "Underwrite Temporal Adj Random", 
                   "Plume Adj", 
                   "Fixed Site Temporal Adj + Plume Adj",  
                   "Fixed Site Temporal Adj Random + Plume Adj",  
                   
                   "Underwrite Temporal Adj + Plume Adj", "Underwrite Temporal Adj Random + Plume Adj")

  visit_levels <- unique(dt$visits) %>% as.numeric() %>% sort()
  visit_levels <- paste(visit_levels, "visits")
  
  dt %>%
    mutate(
      
      visits = str_extract(visits, "\\d+"),
      version = as.character(version),
      version = ifelse(adjusted=="adjusted", paste(version, "plume adj"), version),
      version = gsub("hours", "hr", version),
      version = gsub("business", "bus", version),
      version = str_to_title(version),#) %>% distinct(version)
      
      time_selection = ifelse(grepl("All Hr", version), "All Days & Hours Sampling",
                        ifelse(grepl("Bus Hr", version), "Weekday Business Hours Sampling", NA)),
      time_selection = factor(time_selection, levels=c("Weekday Business Hours Sampling", "All Days & Hours Sampling")),
      version = ifelse(version %in% c("All Hr","Bus Hr"), "No Adjustment", version),
      version = gsub("All Hr |Bus Hr ", "", version),
      version = gsub("Temp ", "Temporal ", version),
      version = gsub("Temporal Adj 1", "Fixed Site Temporal Adj", version),
      version = gsub("Temporal Adj 2", "Underwrite Temporal Adj", version),
         
      version = gsub("Adj Plume Adj", "Adj + Plume Adj", version), 
      version = gsub("Random Plume Adj", "Random + Plume Adj", version), 
      
      version = factor(version, levels= adjustments), 
      
      design = case_when(
        design=="route" ~ "balanced routes",
        design=="balanced" ~ "balanced segments",
        design=="unbalanced" ~ "unbalanced segments",
        design=="random" ~ "unbalanced random clusters",
        design=="sensible" ~ "unbalanced sensible clusters",
        design=="unsensible" ~ "unbalanced unsensible clusters",
        design=="road_type" ~ "unbalanced road type",
        TRUE ~ design
        ),
      
      design = str_to_title(design),
      
      design = gsub("alanced ", "alanced\n", design),
      design = factor(design, levels=gsub("alanced ", "alanced\n", design_levels)),
      
      #design = factor(design, levels = design_levels),
      
      adjusted = str_to_title(adjusted),
      visit_label = paste(visits, "visits"),
      visit_label = factor(visit_label, levels=visit_levels)
      )  
  }

########################################################################################
# default plot
plot_values <- function(dt, 
                        summary_var, y_label,
                        add_shaded_area= FALSE,
                        add_dahsed_line=FALSE,
                        ref_dt, # reference values (e.g., from stationary MM results)
                        grouping_variables = c("design", "version", "time_selection", "visit_label")){
  
  t <- dt %>%
    filter(model != ref_model_coefs$model) %>%
    group_by(across(all_of(grouping_variables))) %>%

    summary_table(var = summary_var) 
  
  p <- t %>%
    
    ggplot(aes(x=version, fill=design)) +
    facet_grid(rows = vars(visit_label), cols = vars(time_selection), switch = "both", scales = "free_x", space="free_x") +
    scale_x_discrete(labels = function(x) str_wrap(x, width = 10))
  
  # for median (IQR)
  if(add_shaded_area==TRUE){
    p <- p + 
      geom_rect(inherit.aes = F, data = ref_dt, aes(xmin=-Inf, xmax=Inf, ymin=q25, ymax=q75), alpha=0.2, fill=ref_model_col) +
      geom_hline(yintercept = ref_dt$median, alpha=0.8, col=ref_model_col, linetype=2)
  }
  
  # for a dashed ref line
  if(add_dahsed_line==TRUE){
    p <- p + geom_hline(yintercept = ref_dt, alpha=0.8, col=ref_model_col, linetype=2)
    }
  
  p <- p + 
    geom_boxplot(position = "dodge", stat="identity", aes(ymin=Q10, lower=Q25, middle=Q50, upper=Q75, ymax=Q90)) +
    labs(x = NULL, 
       y = y_label, 
       fill="Spatial\nBalance",
       linetype=NULL)
  
  return(list(plot=p,
         table=t))
  }



##################################################################################################
# LOAD DATA
##################################################################################################
# raw mobile data
#cov <- readRDS(file.path("data", "onroad", "annie", "cov_onroad_preprocessed.rds")) #%>% pull(native_id)
project_crs <- 4326  #lat/long
m_crs <- 32148

clusters <- readRDS(file.path("data", "onroad", "annie", "segment_clusters_updated.rds")) %>%
  select(-c(latitude, longitude)) %>%
  # drop annie's old clusters for now
  filter(cluster_type != "cluster3")

cov <- readRDS(file.path("data", "onroad", "dr0364d_20230331_modified.rda")) %>%
  st_as_sf(coords = c('longitude', 'latitude'), crs=project_crs, remove = F) %>%
  st_transform(m_crs)  
  
onroad_cov <- cov %>%
  # add cluster IDs
  left_join(clusters, by = c("location" = "id")) %>%
  select(location, cluster_type, #cluster, 
         everything())

saveRDS(onroad_cov, file.path("data", "onroad", "dr0364d_20230331_modified_with_clusters.rda"))

# crosswalks
cw <- read.csv(file.path(dt_path, "onroad", "onroad_model_cw.csv")) %>%
  # drop annie's old clusters for now
  filter(!grepl("_c3_", model)) %>%
  mutate(variable="pnc_noscreen",
         visits = gsub(" visits", "", visits)) %>%
  label_pollutants() 

#cw %>% filter(is.na(version2)) %>% distinct(design, version, version2)

write.csv(cw, file.path(output_final, "onroad_model_cw_clean.csv"), row.names = F)

stationary_cw <- read.csv(file.path(dt_path, "model_cw.csv")) %>%
  filter(variable =="pnc_noscreen",
         design == "full") %>% 
  label_pollutants()  %>%
  mutate(visits = "29",
         adjusted = "Unadjusted")

#model eval
eval_files <- list.files(file.path(dt_path, "onroad", "model_eval")) %>% str_subset(".rda")

model_eval <- lapply(eval_files, function(x){
  readRDS(file.path(dt_path, "onroad", "model_eval", x)) %>%
    # drop annie's old clusters for now
    filter(!grepl("_c3_", model)) %>%
    left_join(select(cw, cluster_type, design, version, visits, adjusted, campaign, model), by="model") 
  }) %>%
  bind_rows()
  
model_eval2 <- add_labels(model_eval)  

write.csv(model_eval2, file.path(output_final, "model_eval_clean.csv"), row.names = F)
#####################################################################

stationary_model_eval <- readRDS(file.path(dt_path, "model_eval.rda")) %>%
  filter(reference == "gs_estimate",
         model %in% stationary_cw$model)  

#ref health estimates from stationary models & extended health analyses
ref_model_coefs <- readRDS(file.path(output_data_path, "model_coefs_extended.rda")) %>%
  filter(model=="s_pncnoscreen_all_01") %>%
  left_join(select(stationary_model_eval, model, reg_based_R2, MSE_based_R2), by="model")
 
tot_pnc <- filter(ref_model_coefs, model=="s_pncnoscreen_all_01") %>% pull(est)

## save as above for reduced health models
ref_model_coefs_reduced <- readRDS(file.path(output_data_path, "model_coefs.rda")) %>%
  filter(model=="s_pncnoscreen_all_01") %>%
  left_join(select(stationary_model_eval, model, reg_based_R2, MSE_based_R2), by="model")
 
tot_pnc_reduced <- filter(ref_model_coefs_reduced, model=="s_pncnoscreen_all_01") %>% pull(est)

############################################################
# person-level exposures 
stationary_cs <- readRDS(file.path(output_data_path, "dt_for_cross_sectional_analysis.rda")) %>%
  rename(exposure = avg_0_5_yr) %>%  
  select(-starts_with("avg_")) %>%
  right_join(stationary_cw, by="model")

# LARGE FILES
model_covars_extended <- readRDS(file.path(output_data_path, "model_covars_extended.rda"))

cs0 <- readRDS(file.path(output_data_path, "dt_for_cross_sectional_analysis_road.rda")) %>%
  select(all_of(c(model_covars_extended, "avg_0_5_yr", "study_id", "casi_irt", "model"))) 
# cs0 <-cs
cs <- cs0 %>%
  right_join(cw, by="model")  

rm(cs0)

# --> error w/ vector memory 

cs <- cs %>% rename(exposure = avg_0_5_yr)  #%>% bind_rows(stationary_cs)


############################################################
# health effect estimates & CW
model_coefs <- readRDS(file.path(output_data_path, "model_coefs_road_extended.rda")) %>%
  right_join(cw, by="model") %>%
  #filter(model %in% cw$model) %>%
  mutate(variable="pnc_noscreen")  

saveRDS(model_coefs, file.path(output_final, "model_coefs.rda"))

## alt - reduced adjustment models
# health effect estimates & CW
model_coefs_reduced_model <- readRDS(file.path(output_data_path, "models_r_all.rda")) %>%
  filter(term=="avg_0_5_yr") %>%
  right_join(cw, by="model") %>%
  mutate(variable="pnc_noscreen")  

saveRDS(model_coefs_reduced_model, file.path(output_final, "model_coefs_reduced_model.rda"))

# onroad model predictions at stationary locations
stationary_prediction_files <- list.files(file.path(dt_path, "onroad", "model_eval", "predictions_at_stationary_sites"))

# x=stationary_prediction_files[1]
onroad_predictions <- lapply(stationary_prediction_files, function(x){
  readRDS(file.path(dt_path, "onroad", "model_eval", "predictions_at_stationary_sites", x))}) %>%
  bind_rows() %>% 
  right_join(cw, by="model")

# site predictions
all_predictions <- readRDS(file.path(dt_path, "UK Predictions", "all_predictions.rda")) %>%
  filter(design == "full",
         variable == "pnc_noscreen") %>%
  mutate(model = "s_pncnoscreen_all_01",
         visits = "29", #as.character(visits)
         version = "All Hr") %>%
  left_join(select(stationary_cw, variable, design, model))  

all_predictions <- bind_rows(onroad_predictions, all_predictions)  

ref_prediction_distribution <- all_predictions %>%
  filter(model == ref_model_coefs$model) %>%
  summarize(n=n(),
            median = median(prediction),
            q25 = quantile(prediction, 0.25),
            q75 = quantile(prediction, 0.75)) 

# reference P-TRAK data - number of visits (stops) per site; for the HEI dataset (not necessarily the "true" health models)
stops_w <- readRDS(file.path(dt_path, "stops_used.rda"))

# GIS
monitoring_area <- readRDS(file.path("data", "GIS", "monitoring_area_shp.rda"))

##################################################################################################
# COMMON VARIABLES
##################################################################################################
#main_designs <- design_levels 
main_designs <- c("Balanced Routes",
                  "Balanced Segments" ,
                  "Unbalanced Random Clusters",
                  "Unbalanced Road Type",
                  
                  "Balanced\nRoutes",
                  "Balanced\nSegments" ,
                  "Unbalanced\nRandom Clusters",
                  "Unbalanced\nRoad Type"
                  )


version_levels <- levels(model_eval2$version) #%>% str_subset(., "Random", negate=T)
saveRDS(version_levels, file.path(output_final, "version_levels.rda"))


# drop random temporal adjustments
main_versions <- version_levels %>% str_subset(., "Random", negate=T)

main_visits <- c(12)
alt_visits <- c(4)

main_cluster <- "cluster1"

reference_stationary_model <- "s_pncnoscreen_all_01"

# UW colors
uw_purple <- "#4b2e83"
uw_gold1 <- "#b7a57a" 
uw_gold2 <- "#85754d"

ref_model_col <- config$ref_model_col  
annotation_line_size <-1.5

################################################
# HEI
hei_designs <- c(#"Balanced Routes", 
                 "Balanced Segments",
                 "Unbalanced Random Clusters", 
                 "Unbalanced Sensible Clusters",
                 "Unbalanced Unsensible Clusters",
                 "Unbalanced Road Type",
                 
                 #"Balanced\nRoutes", 
                 "Balanced\nSegments",
                 "Unbalanced\nRandom Clusters", 
                 "Unbalanced\nSensible Clusters",
                 "Unbalanced\nUnsensible Clusters",
                 "Unbalanced\nRoad Type")

```

```{r}
# visit data (routes vs other)
visit_path <- file.path("data", "onroad", "annie", "v2", "visits")
file_names <- list.files(visit_path) %>%
    grep("route_unadjusted_12_balanced|nonspatial_unadjusted_12_balanced", value = T, .)

visits <- lapply(file_names, function(f){readRDS(file.path(visit_path, f))}) %>% bind_rows() %>%
  left_join(select(stationary_cw, variable, design, model))  
  

#fixed_site_adjusted_visits <- readRDS(file.path(dt_path, "temporal_adjustment.rda"))
```

```{r}
# temporal adjustments, hourly
fixed_site_adj <- readRDS(file.path("data", "epa_data_mart", "wa_county_nox_temp_adjustment.rda")) %>%
  select(time, hourly_adjustment = diff_adjustment_winsorize_ptrak) %>%
  mutate(adjustment = "Fixed Site")

main_bg <- "hr3_pct1" 

underwrite_adj <- readRDS(file.path("data", "onroad", "annie", "v2", "temporal_adj", "20240421", "underwrite_temp_adj_no_hwy.rda")) %>%
  filter(background_adj == main_bg) %>%
  select(time, hourly_adjustment = avg_hourly_adj) %>%
  mutate(adjustment = "Underwrite")

hourly_adjustments <- rbind(fixed_site_adj, underwrite_adj)


```



# reference stationary P-TRAK data

```{r}

stops_w %>%
  drop_na(pnc_noscreen) %>%
  group_by(location) %>%
  summarize(visits = n()) %>%
  summarize(
    min = min(visits),
    q25 = quantile(visits, 0.25),
    mean = mean(visits),
    q50 = quantile(visits, 0.50),
    max=max(visits)
  )%>%
  kable(caption = "distribution of visits per stop with unscreened P-TRAK readings") %>%
  kable_styling()

```


# model counts

```{r}
cw %>%
  filter(
    !cluster_type %in% c("cluster2", "cluster3"),
    !grepl("Random", version, ignore.case=T),
    visits %in% c(4, 12)
  ) %>% #filter(campaign==first(campaign)) %>% View()
  #prep_for_figs() %>%
  distinct(design, version, visits, cluster_type) %>% 
     
  kable() %>%
  kable_styling()

# 112*30 #3360

```

# map

```{r}
# Define the bounding box for King County
bbox <- st_bbox(monitoring_area %>% st_buffer(10) )

# # Convert the bbox to a format suitable for osmdata
# bbox_vector <- c(bbox["xmin"], bbox["ymin"], bbox["xmax"], bbox["ymax"])

# Convert the bbox to a numeric vector suitable for osmdata and ggmap
bbox_vector <- c(as.numeric(bbox["xmin"]),
                 as.numeric(bbox["ymin"]),
                 as.numeric(bbox["xmax"]),
                 as.numeric(bbox["ymax"]))

# Create the query for major roads and get the data in sf format
major_roads <- osmdata::opq(bbox_vector) %>%
  add_osm_feature(key = "highway", value = c(#"motorway", 
    #"trunk", 
                                             "primary"#, "secondary"
                                             )) %>%
  osmdata_sf() %>%
  .$osm_lines

#plot(major_roads[1])

##########################################
# x=unique(onroad_cov$cluster_type)[1]
lapply(unique(onroad_cov$cluster_type), function(x){
  p <- onroad_cov %>%
    filter(cluster_type==x) %>%
    mutate(cluster_type = case_when(
      cluster_type == "cluster1" ~ "Main Clusters",
      cluster_type == "cluster2" ~ "Sensitivity (Larger) Clusters",
    )) %>%
    
    ggplot(aes(col=cluster_value), ) +
    facet_wrap(~cluster_type) + 
    annotation_map_tile(type = "osm", zoom = 11, alpha = 0.5) +  # Add OpenStreetMap tiles as background

    # roads
    geom_sf(data = major_roads, color = "gray", alpha=0.3) +
    # clusters
    geom_sf(size=0.5, show.legend=F) +
  
    # add scale & N arrow to top rught
    annotation_scale(location = "tr") +
    annotation_scale(location = "tr", unit_category ="imperial", pad_y = unit(0.55, "cm")) +
    annotation_north_arrow(location = "tr", pad_y = unit(0.5, "in"), style = north_arrow_fancy_orienteering) +
    
    theme_bw() +
    #theme_minimal() +
    theme(# remove the vertical grid lines
          panel.grid.major = element_blank()) + 
    scale_color_viridis_c() #+ labs(caption = "Map data © OpenStreetMap contributors")
  
  # for HEI report etc.
  ggsave(plot = p, file.path(image_path, "SI", paste0(x, ".png")), width = 5, height = 8)
  
  p
  
  }) %>%
  ggarrange(plotlist = .) %>%
  annotate_figure(bottom = text_grob("Map data © OpenStreetMap contributors",
                                     hjust = 1, 
                                     x=1))

ggsave(file.path(image_path, "SI", "cluster_map.png"), width = 8, height = 8)

```

# visits (route vs other)

more sampling days
```{r}
visit_count <- visits %>%
  group_by(design, version, campaign, #visits, adjusted, #campaign
           ) %>%
  mutate(version = gsub("by route ","", version)) %>%
  
  summarize(
    n=n(),
    days = length(unique(date)),
    #mean = mean(median_value),
    sd = sd(median_value),
    #range = paste(round(range(median_value)), collapse=", ")
  ) 

visit_count %>% 
  group_by(design, version) %>%
  summarize(
    mean_visits = mean(n),
    campaigns = length(unique(campaign)),
    mean_days = mean(days),
    mean_sd = mean(sd)
    
  ) %>%
  
  arrange(version)


# use acf()?


```

# temporal adjustments

distribution table 

```{r}
hourly_adjustments %>%
  group_by(Adjustment=adjustment) %>%
  #summary_table(var = "hourly_adjustment") %>%
  summarize(
    Hours = n(),
    Min = min(hourly_adjustment),
    #Q10 = quantile(hourly_adjustment, 0.10),
    Q25 = quantile(hourly_adjustment, 0.25),
    Median = quantile(hourly_adjustment, 0.50),
    Mean = mean(hourly_adjustment),

    Q75 = quantile(hourly_adjustment, 0.75),
    #Q90 = quantile(hourly_adjustment, 0.90),
    Max = max(hourly_adjustment),
   
    SD = sd(hourly_adjustment),
    IQR = IQR(hourly_adjustment)#,
    #IQR = paste(round(Q25), ", ", round(Q75)),
    #'Q10-90 Range' = Q90-Q10
    #Range = round(Max-Min)
  ) %>%
  #select(-c(Q25, Q75)) %>%
  kable(caption = "Distribution of hourly adjustments from temporal adjustment approaches", 
        digits = 0, format.args = list(big.mark = ",")) %>%
  kable_styling()
  


```



# Annual average estimates

```{r}
# table/boxplots of distribution
estimates <- all_predictions %>%
  filter(grepl("^s_", model)) %>%
  distinct(location, gs_estimate, prediction) %>% 
  pivot_longer(cols = c(gs_estimate, prediction)) %>% 
  group_by(name) %>%
  summary_table(var = "value") 

estimates %>%
  kable(caption = "Distribution of estimated and all-data stationary model predicted annual average PNC at 309 stationary locations", 
      digits = 0,
      format.args = list(big.mark=",")
      ) %>%
kable_styling()

```

## HEI

```{r}
r1 <- all_predictions %>%
  add_labels() %>%
  keep_hei_results() %>%
  
  plot_values(summary_var = "prediction", 
              y_label = "Predicted PNC (pt/cm3)",
              add_shaded_area = T, 
              ref_dt=ref_prediction_distribution)

r1$plot 
     
ggsave(file.path(image_path, "HEI Report", "Appendix", "predicted_ufp_309_sites.png"), width=10, height=6)

```


```{r}
print("dashed line is median (IQR) predicted PNC from the all-data stationary model")

r2 <- all_predictions %>%
  add_labels() %>% 
  keep_main_results() %>%
  plot_values(summary_var = "prediction",
              y_label = "Predicted PNC (pt/cm3)",
              add_shaded_area = T, ref_dt=ref_prediction_distribution)

r2$plot 

ggsave(file.path(image_path, "SI", "predicted_ufp_309_sites.png"), width=8, height=8)

```


# Model R2

model eval is at 309 stationary sites


results           
* BH unadjusted R2 are all 0     
* All hours unadju are all 0 except for 1 value

note: results are different now  for the 4-visit results for clustered analyses. things look more as expected w/ upated clusters.

```{r, fig.height=8, fig.with=8}
stationary_r2_mse_ref <- ref_model_coefs$MSE_based_R2

print(paste("reference model R2 is the all-data stationary model:", round(stationary_r2_mse_ref, 2)))

r3 <- model_eval %>%
  add_labels() %>%  
  keep_main_results() %>%  

  plot_values(summary_var = "MSE_based_R2", 
              y_label = expression("Exposure Model" ~ R[MSE]^2),
              add_dahsed_line = T, ref_dt = stationary_r2_mse_ref)

r3$plot + 
  theme(text = element_text(size=12))


ggsave(file.path(image_path, "r2_mse.png"), width = 8, height = 6)

 # save numbers
r3$table %>%
  mutate_if(is.numeric, ~round(., 2)) %>%
  write.csv(file.path(image_path, "r2_mse.csv"), row.names = F)

#############################################
# SI: cluster comparison

cluster_designs <- str_subset(design_levels, "Cluster|Road Type") %>%
  gsub("alanced ", "alanced\n",.)

model_eval %>%
  add_labels() %>%
  filter(visits %in% 12,
         design %in% cluster_designs,
         !grepl("Sensible|Unsensible", design),
         cluster_type %in% c("cluster1", "cluster2"),
         !grepl("Random", version)) %>%
  drop_na(cluster_type) %>%
  mutate(cluster_type = ifelse(cluster_type=="cluster1", "Main Clusters",
                               ifelse(cluster_type=="cluster2", "Sensitivity (Larger) Clusters", NA))) %>%
  
  
  group_by(across(all_of(c("design", "version", "time_selection", "visit_label", "cluster_type")))) %>%
  summary_table(var = "MSE_based_R2") %>%
  
  ggplot(aes(x=version, fill=design)) +
  facet_grid(rows = vars(cluster_type, visit_label), cols = vars(time_selection), switch = "both", scales = "free_x", space="free_x")  + geom_hline(yintercept = stationary_r2_mse_ref, alpha=0.8, col=ref_model_col, linetype=2) +  
    
  geom_boxplot(position = "dodge", stat="identity", aes(ymin=Q10, lower=Q25, middle=Q50, upper=Q75, ymax=Q90)) +
  scale_x_discrete(labels = function(x) str_wrap(x, width = 10)) +
  labs(x = NULL, 
       y = expression("Exposure Model" ~ R[MSE]^2), 
       fill=NULL)
  
ggsave(file.path(image_path, "SI", paste0("r2_mse_clusters.png")), width = 8, height = 6)


#############################################
# SI: routes - increasing visits
model_eval %>%
  add_labels() %>%
  keep_main_results(designs_to_keep = c("Balanced\nRoutes", "Balanced\nSegments"), visits_to_keep = c(4, 20)) %>%
  filter(model != ref_model_coefs$model) %>%
  group_by(across(all_of(c("design", "version", "time_selection", "visit_label")))) %>%
  summary_table(var = "MSE_based_R2") %>%
    
  ggplot(aes(x=version, fill=design, linetype=visit_label)) +
  facet_grid(cols = vars(time_selection), switch = "both", scales = "free_x", space="free_x") +
  scale_x_discrete(labels = function(x) str_wrap(x, width = 10)) +
  geom_hline(yintercept = stationary_r2_mse_ref, alpha=0.8, col=ref_model_col, linetype=2) +
  geom_boxplot(position = "dodge", stat="identity", aes(ymin=Q10, lower=Q25, middle=Q50, upper=Q75, ymax=Q90)) +
  
  scale_linetype_manual(values = c(#3,
                                   2,1)) +
    
  labs(x = NULL, 
       y = expression("Exposure Model" ~ R[MSE]^2), 
       fill="Spatial\nBalance",
       linetype="Visits"
       )
    
ggsave(file.path(image_path, "SI", paste0("r2_mse_routes.png")), width = 8, height = 6)

#############################################
```


```{r, fig.height=8, fig.with=8}
# random temporal adjustment
random_versions <- version_levels %>% 
  str_subset(., "Temporal Adj|No Adjustment") %>%
  str_subset(., "Plume", negate=T)

df <- model_eval %>%
  add_labels() %>%  
  keep_main_results(versions_to_keep = random_versions, 
                    times_selections_to_keep = "Weekday Business Hours Sampling", 
                    visits_to_keep = "12"
                    ) %>%  
  #filter(!grepl("Road Type", design)) %>%
  
  # plot_values(summary_var = "MSE_based_R2", y_label = expression("Exposure Model" ~ R[MSE]^2),
  #             add_dahsed_line = T, ref_dt = stationary_r2_mse_ref)

select(#model, 
       MSE_based_R2, design, version, campaign, visit_label) 

# Separate the "No Adjustment" and other versions
df_no_adjustment <- df %>% 
  filter(version == "No Adjustment") %>%
  select(MSE_based_R2, design, campaign, visit_label) %>%
  rename(MSE_based_R2_no_adjustment = MSE_based_R2)

df_other_versions <- df %>% 
  filter(version != "No Adjustment") %>%
  #select(model, MSE_based_R2, design, visit_label) %>%
  rename(MSE_based_R2_other = MSE_based_R2)

# Join the two datasets by 'model', 'design', and 'visit_label'
df_merged <- df_no_adjustment %>%
  inner_join(df_other_versions, by = c("design", "campaign", "visit_label")) %>%
  mutate(temporal_adjustment = ifelse(grepl("Fixed Site", version), "Fixed Site Temporal Adjustment",
                                      ifelse(grepl("Underwrite", version), "Underwrite Temporal Adjustment", NA)),
         random_adjustment = ifelse(grepl("Random", version), "Random", "Main"))

df_merged %>%
  # drop random noise
  filter(random_adjustment=="Main") %>%
  
  ggplot(aes(x=MSE_based_R2_no_adjustment, y= MSE_based_R2_other, col=design, 
             #shape=random_adjustment, linetype=random_adjustment
             )) + 
  facet_grid_equal(#rows = vars(visit_label),
                   cols=vars(temporal_adjustment), 
                   #switch = "both"
                   ) + 
  geom_abline(slope = 1, intercept = 0, linetype=2) +
  geom_point(alpha=0.4) + 
  geom_smooth(se=F) +
  scale_x_continuous(limits = c(0, 0.4)) +
  scale_linetype_manual(values=c(1,4))+
  labs(x= expression(R[MSE]^2 ~ "Without Temporal Adjustment"),
       y= expression(R[MSE]^2 ~ "With Temporal Adjustment"),
       # linetype="Temporal\nAdjustment",
       # shape="Temporal\nAdjustment",
       col="Design"
       ) + 
  theme(aspect.ratio = 1)

ggsave(file.path(image_path, "SI", "r2_random_temporal_adj.png"), width = 8, height = 6)

```

regression-based R2 comparison 

```{r}
#############################################
print(paste("dashed line is the MSE-based R2 from the all-data stationary model. both MSE and reg-based R2 are 0.77")) 

compare_r2_plot <- function(dt){
  dt %>%
  pivot_longer(cols = contains("R2")) %>%
  group_by(across(all_of(c("design", "version", "time_selection", "visit_label", "name")))) %>%
  summary_table(var = "value") %>%
  mutate(name = ifelse(name=="MSE_based_R2", "MSE-based R2 (Main)", "Regression-based R2")) %>%
  
  ggplot(aes(x=version, fill=design)) +
  #facet_grid(rows = vars(time_selection, visit_label), cols = vars(name), switch = "both", scales = "free_x", space="free_x")  + 
    facet_grid(rows = vars(time_selection, visit_label),
               cols = vars(name),
               switch = "both", scales = "free_x", space="free_x") +
      
    
    geom_hline(yintercept = ref_model_coefs$MSE_based_R2, alpha=0.8, col=ref_model_col, linetype=2) +  
  geom_boxplot(position = "dodge", stat="identity", aes(ymin=Q10, lower=Q25, middle=Q50, upper=Q75, ymax=Q90)) +
    scale_x_discrete(labels = function(x) str_wrap(x, width = 10)) +
    labs(x = NULL,
         y = expression("Exposure Model" ~ R[Regression]^2),
         fill="Spatial\nBalance")
  
}

# see similar patterns for 4 visits
model_eval %>%
  add_labels() %>% 
  keep_main_results(visits_to_keep = 12) %>%
  #compare_r2_plot() 
  
  pivot_longer(cols = contains("R2")) %>%
  group_by(across(all_of(c("design", "version", "time_selection", "visit_label", "name")))) %>%
  summary_table(var = "value") %>%
  mutate(name = ifelse(name=="MSE_based_R2", "MSE-based\nR2 (Main)", "Regression-based\nR2")) %>%
  
  ggplot(aes(x=version, fill=design, linetype=name, col=name)) +
  
  facet_grid(rows = vars(visit_label),
               cols = vars(time_selection), #vars(name),
               switch = "both", scales = "free_x", space="free_x") +
  geom_hline(yintercept = ref_model_coefs$MSE_based_R2, alpha=0.8, col=ref_model_col, linetype=2) +  
  geom_boxplot(position = "dodge", stat="identity", aes(ymin=Q10, lower=Q25, middle=Q50, upper=Q75, ymax=Q90)) +
  scale_x_discrete(labels = function(x) str_wrap(x, width = 10)) +
  
  scale_color_manual(values=c("black", uw_gold1)) +
  
  
  labs(x = NULL,
       y = expression("Exposure Model" ~ R^2 ~ "(MSE or Regression)"),
       fill="Spatial\nBalance",
       col="R2",
       linetype="R2")
  
ggsave(file.path(image_path, "SI", "r2_regression.png"), width = 9, height = 6)

```

## HEI Report

* temp adj makes R2 values noisier; makes plume adjustments noiser

```{r}

r4 <- model_eval %>%
  add_labels() %>%
  keep_hei_results() %>%
  plot_values(summary_var = "MSE_based_R2", y_label = expression("Exposure Model" ~ R[MSE]^2),
              add_dahsed_line = T, ref_dt = stationary_r2_mse_ref)
  
r4$plot

ggsave(file.path(image_path, "HEI Report", paste0("r2_mse.png")), width = 10, height = 8)
   
# save table
r4$table %>%
  select(design, version, visits=visit_label, time_selection, N, Min, Q10, Q25, Q50, Q75, Q90, Max) %>%
  mutate_if(is.numeric, ~round(., 1)) %>% 
  write.csv(file.path(image_path, "HEI Report", "r2_mse.csv"), row.names = F)


 # SI - regression R2 (formatted differently)
model_eval %>%
  add_labels() %>% 
  keep_hei_results(keep_visits = 12) %>%
  compare_r2_plot() 
  
ggsave(file.path(image_path, "HEI Report", "Appendix", "r2_regression.png"), width = 10, height = 8)

```


lms 

```{r, eval=F}

fit_r2_lm <- function(y="MSE_based_R2",
                      x = c("design", "version", "plumed_adjusted", "visits")
                      ){
  
  lm_formula <- paste(y, "~", paste(x, collapse = "+"))
  
  model_eval %>%
  filter(!grepl("Random", version),
         visits %in% c(4,12)) %>%
  mutate(
    version = gsub(" Plume Adj", "",  version),
    design = factor(design, levels=rev(design_levels)),
    plumed_adjusted = factor(adjusted, levels=c("Unadjusted", "Adjusted"))) %>%
  
  lm(as.formula(lm_formula), data=.)

}


# units of analysis are the campaign-specific R2 (30 per design-version-etc.)
# mse-based R2 ~ design_features
fit_r2_lm() %>% 
  broom::tidy()

# reg-R2
fit_r2_lm(y = "reg_based_R2") %>% 
  broom::tidy()

```


# Participant Exposrue Predictions 

comparison of predicted 5 yr exposures at baseline

```{r, fig.height=8, fig.with=10}
# reference design
stationary_cs %>%
  summarize(
    N = n(),
    Participants = length(unique(study_id)),
    Campaigns = length(unique(campaign)),
    Min = min(exposure),
    Q05 = quantile(exposure, 0.05),
    Q25 = quantile(exposure, 0.25),
    Q50 = quantile(exposure, 0.50),
    Q75 = quantile(exposure, 0.75),
    Q95 = quantile(exposure, 0.95),
    Max = max(exposure)) %>%
  kable(caption = "Distribution of Predicted UFP (pt/cm3) exposure from the reference All data stationary exposure model", digits = 0, format.args = list(big.mark=",")) %>%
  kable_styling()

# cohort predictions from ref model
ref_cs_distribtuion <- stationary_cs  %>%
  summarize(n = n(),
            median = median(exposure),
            q25 = quantile(exposure, 0.25),
            q75 = quantile(exposure, 0.75))  
# plot
cohort_exposures <- cs %>%
  add_labels() %>%  
  keep_main_results() 

r5 <- cohort_exposures %>%
  plot_values(summary_var = "exposure", 
              y_label = "Predicted PNC (pt/cm3)",
              ref_dt = ref_cs_distribtuion,
              add_shaded_area= TRUE#, add_dahsed_line=TRUE
              )

r5$plot

ggsave(file.path(image_path, "SI", "cohort_exposures.png"), width = 8, height = 8)

```

## HEI

```{r}
r6 <- cs %>%
  add_labels() %>%
  keep_hei_results() %>%
  plot_values(summary_var = "exposure", 
              y_label = "Predicted PNC (pt/cm3)",
              ref_dt = ref_cs_distribtuion)

r6$plot

ggsave(file.path(image_path, "HEI Report", "Appendix", "cohort_exposures.png"), width = 8, height = 8)

```



# Health effect estimates (CASI-IRT)

```{r}
ref_model_coefs %>%
  select(model, est, lower, upper, MSE_based_R2, reg_based_R2) %>%
  kable(caption = "health estimate from the reference all-data stationary exposure model", 
        digits = 3) %>%
  kable_styling()

ref_model_coefs_reduced %>%
  select(model, est, lower, upper, MSE_based_R2, reg_based_R2) %>%
  kable(caption = "REDUCED health estimate from the reference all-data stationary exposure model", 
        digits = 3) %>%
  kable_styling()

```

```{r, fig.height=10, fig.with=10}
# function to produce main health results figures
# dt= model_coefs_reduced_model %>%
#     add_labels() %>%
#     keep_main_results(visits_to_keep = v) %>%
#     mutate(visit_label = paste0("Reduced Health Model (", visit_label, ")"))

# reference <- ref_model_coefs_reduced$est
# lower <- ref_model_coefs_reduced$lower
# upper <- ref_model_coefs_reduced$upper

health_plot <- function(dt,
                        reference, reference_text_offset=0,
                        lower, upper,
                        null_text_offset=0
                        ){
  
  p <- dt %>%
    group_by(design, time_selection, version, visit_label) %>%
    summary_table(var = "est") %>%
    
    ggplot(aes(x=version, fill=design,)) +
    facet_grid(rows = vars(visit_label), cols = vars(time_selection),
                 switch = "both", scales="free_x",space="free") +
    
    annotate("rect", xmin = -Inf, xmax = Inf, ymin = lower, ymax = upper, 
           alpha = 0.1, fill = ref_model_col) +
 
    geom_hline(yintercept = reference, linetype=2, col=ref_model_col,) +
    geom_hline(yintercept = 0, linetype=2, col="red") +
  
    annotate("text", x = Inf, y = reference + reference_text_offset, label = "Reference", 
         hjust = 1, vjust = -0.5, color = ref_model_col, size = 3.5) +
   annotate("text", x = Inf, y = 0+null_text_offset, label = "No Association", 
         hjust = 1, vjust = -0.5, color = "red", size = 3.5) +
    
    geom_boxplot(position = "dodge", stat="identity", aes(ymin=Q10, lower=Q25, middle=Q50, upper=Q75, ymax=Q90)) +
    scale_y_continuous(sec.axis = sec_axis(trans=~(.-reference)/-reference, labels = scales::percent), 
                       position="left", 
                       limits = c(config$health_plot$min_y_ci, config$health_plot$max_y_ci)
                         ) +
    scale_x_discrete(labels = function(x) str_wrap(x, width = 8)) +  
    
    #scale_fill_viridis_d() +
    #scale_fill_manual(values = hei_cb_palette_designs) +
    #scale_fill_brewer(palette = "Set2") + #set2 is cb friendly
    #scale_fill_manual(values = cb_palette) +
      
    theme(panel.spacing = unit(0.2, "lines")) +
    
    labs(y=NULL, 
         fill="Spatial\nBalance",
         x = NULL, linetype=NULL) 

  p
  
  # p + scale_fill_manual(values = hei_cb_palette_designs) +
  # theme_minimal() + # A minimal theme to reduce clutter
  # theme(legend.position = "bottom",  # Place legend at bottom for consistency
  #       panel.spacing = unit(2, "lines"))  # Add spacing between panels
  
}

```

save table numbers 

```{r}
# reduced models
model_coefs_reduced_model_table <- model_coefs_reduced_model %>%
  add_labels() %>%
  keep_main_results() %>%
  mutate(design = gsub("\n", " ", design)) %>%
  group_by(design, time_selection, version, visit_label) %>%
  summary_table(var = "est") %>%
  ungroup()

model_coefs_reduced_model_table %>%
  mutate_if(is.numeric, ~round(., 3)) %>%
  write.csv(., file.path(image_path, "betas_reduced_health_model.csv"), row.names = F)


# extended models
model_coefs_extended_model_table <- model_coefs %>%
  add_labels() %>%
  keep_main_results() %>%
  mutate(design = gsub("\n", " ", design)) %>%
  group_by(design, time_selection, version, visit_label) %>%
  summary_table(var = "est") %>%
  ungroup()

model_coefs_extended_model_table %>%
  mutate_if(is.numeric, ~round(., 3)) %>%
  write.csv(., file.path(image_path, "betas_extended_health_model.csv"), row.names = F)

```


```{r, fig.height=10, fig.with=10}

# v=12
lapply(c(4, 12), function(v){
  
  reduced_health_model <- model_coefs_reduced_model %>%
    add_labels() %>%
    keep_main_results(visits_to_keep = v) %>%
    mutate(visit_label = paste0("Reduced Health Model (", visit_label, ")")) %>%
    health_plot(dt=., reference_text_offset = -0.005,
                reference = ref_model_coefs_reduced$est,  
                lower=ref_model_coefs_reduced$lower,
                upper = ref_model_coefs_reduced$upper) + 
    theme(text = element_text(size=16),
          #axis.text.x = element_text(size=16)
          )
  
  extended_health_model <- model_coefs %>%
    add_labels() %>%
    keep_main_results(visits_to_keep = v) %>%
    mutate(visit_label = paste0("Primary Health Model (", visit_label, ")")) %>%
    health_plot(dt=., 
                reference = ref_model_coefs$est, 
                lower=ref_model_coefs$lower,
                upper = ref_model_coefs$upper, 
                null_text_offset = -0.007) + 
    theme(text = element_text(size=16))

  ggarrange(reduced_health_model, extended_health_model, ncol = 1, 
          common.legend = T, legend = "bottom") %>%
  annotate_figure(left = text_grob(expression(beta[m]), 
                                 rot = 90, size = 16),
                right = text_grob("% Difference from Reference", 
                                  rot = 270, size = 16))


if(v==12){ggsave(file.path(image_path, "betas.png"), bg="white", width = 11, height = 11)} 

if(v==4){ggsave(file.path(image_path, "SI", paste0("betas_", v, "_visits.png")), bg="white", width = 11, height = 11)}

})

```


```{r, eval=F}
# routes
model_coefs %>%
  add_labels() %>% 
  keep_main_results(designs_to_keep = c("Balanced\nRoutes", "Balanced\nSegments"), 
                    visits_to_keep = c(4, 12, 20)) %>%
  
  health_plot(dt=.,
              reference_text_offset = 0,
              reference = ref_model_coefs$est,  
              lower=ref_model_coefs$lower,
              upper = ref_model_coefs$upper, 
              null_text_offset = -0.005
              ) + 
  labs(y = "Health Point Estimate") + 
  # add second y-axis label to this plot
  scale_y_continuous(sec.axis = sec_axis(trans=~(.-ref_model_coefs$est)/-ref_model_coefs$est,
                                         name= "Health Point Estimate % Difference",
                                         labels = scales::percent),
                     position="left") 
  
ggsave(file.path(image_path, "SI", "betas_routes.png"), width = 8, height = 6)

```

## HEI Report 

```{r}
# v=12
lapply(c(4, 12), function(v){
  
  reduced_health_model <- model_coefs_reduced_model %>%
    add_labels() %>%
    keep_hei_results(keep_visits = v) %>%
    mutate(visit_label = paste0("Health Model 1 (", visit_label, ")")) %>%
    health_plot(dt=., reference_text_offset = -0.005,
                reference = ref_model_coefs_reduced$est,  
                lower=ref_model_coefs_reduced$lower,
                upper = ref_model_coefs_reduced$upper)  

  
  extended_health_model <- model_coefs %>%
    add_labels() %>%
    keep_hei_results(keep_visits = v) %>%
    mutate(visit_label = paste0("Health Model 2 (", visit_label, ")")) %>%
    health_plot(dt=., 
                reference = ref_model_coefs$est, 
                lower=ref_model_coefs$lower,
                upper = ref_model_coefs$upper, 
                null_text_offset = -0.004
                )  
  
  ggarrange(reduced_health_model, extended_health_model,
            ncol = 2, 
          common.legend = TRUE, 
          legend = "bottom") %>%
  annotate_figure(left = text_grob(expression(beta), 
                                   rot = 90, size = 14),
                right = text_grob("% Difference", 
                                  rot = 270, size = 14))
    
  
if(v==12){ggsave(file.path(image_path, "HEI Report", "betas.png"), 
                 bg = "white",
                 width = 12, height = 6
                 )} 

  if(v==4){ggsave(file.path(image_path, "HEI Report", "Appendix", paste0("betas_", v, "_visits.png")), 
                  bg = "white",
                  width = 12, height = 6
                  )}

})

```

## Point & 95% CI


```{r}
rectangle_alpha <- 0.15

plot_point_ci <- function(beta_data, reference_data){
  
  beta_data %>%
  add_labels() %>%
  keep_main_results(visits_to_keep = 12,
                    versions_to_keep = c("No Adjustment",
                                          "Plume Adj"#,
                                         #"Underwrite Temporal Adj"#,
                                         #"Fixed Site Temporal Adj"
                                         )

                    ) %>% 
   mutate(design=gsub("\n", " ", design)) %>%
  
  ggplot(aes(x=campaign, shape=version, col=version)) +
  facet_wrap(~design+time_selection, ncol = 2) + 
  
  geom_hline(yintercept = 0, linetype=2, alpha=0.2, col="red") +
  geom_rect(inherit.aes = F, data=reference_data, aes(ymin = lower, ymax = upper, xmin=-Inf, xmax=Inf), alpha=rectangle_alpha, fill=ref_model_col) +
  geom_hline(data=reference_data, linetype=2, aes(yintercept = est), col=ref_model_col) +
  geom_pointrange(aes(y=est, ymin=lower, ymax=upper), size=0.4,
                  position=position_dodge(width=0.7)) +
  scale_x_continuous(breaks = c(1, seq(10,30,10))) +
  labs(x="Campaign", 
       y= expression(beta[m] ~ " (95% CI)"),
       #y = "Health Estimate (95% CI)", 
       shape=NULL, col=NULL)
  
  
}
 
# reduced model
plot_point_ci(model_coefs_reduced_model, ref_model_coefs_reduced)

ggsave(file.path(image_path, "SI", "beta_point_ci_reduced.png"), width = 8, height = 10)

# extended model
plot_point_ci(model_coefs, ref_model_coefs)

ggsave(file.path(image_path, "SI", "beta_point_ci_extended.png"), width = 8, height = 10)
       

```
