==> R - hei_aim3a/9_stationary_summary.Rmd <==
---
title: "Epi results using stationary data exposure models"
author: "Magali Blanco"
date: ' `r Sys.time()` '
output:
  html_document:
    df_print: paged
    toc: yes
    toc_depth: '3'
    number_sections: true
    toc_float: true
    collapsed: false
    smooth_scroll: false
editor_options: 
  chunk_output_type: console
---

# Setup 

```{r setup, include=FALSE}
# Rscript -e 'rmarkdown::render("9_stationary_summary.Rmd", "html_document")'
knitr::opts_chunk$set(echo = F, 
                      cache=F, cache.comments = F, 
                      message = F, warning = F, 
                      tidy.opts=list(width.cutoff=60), tidy=TRUE#, fig.height = 10, fig.width = 10
                      )  

# Clear workspace of all objects and unload all extra (non-base) packages
rm(list = ls(all = TRUE))
if (!is.null(sessionInfo()$otherPkgs)) {
  res <- suppressWarnings(
    lapply(paste('package:', names(sessionInfo()$otherPkgs), sep=""),
           detach, character.only=TRUE, unload=TRUE, force=TRUE))}

pacman::p_load(tidyverse, 
               PRISMAstatement,table1,
               knitr, kableExtra,
               ggpubr, #ggarrange()
               ggmap, sf, #mapping
               ggspatial, #mapping, adding scales, N arrows...
               spData, # us_states data - mapping WA state
               cowplot #layering ggplots. ggdraw(), draw_plot()
               )    

config <- yaml::read_yaml("config.yaml")

dt_path <- file.path("Output", readRDS(file.path("Output", "latest_dt_version.rda")))

set.seed(1)

source("functions.R")
output_data_path <- file.path(dt_path, "epi", "20240725")

image_path <- file.path("..", "Manuscript", "Images", "v5") # "v4")

if(!dir.exists(file.path(image_path, "other", "presentation"))) {dir.create(file.path(image_path, "other", "presentation"), recursive = T)}
if(!dir.exists(file.path(image_path, "other", "poster"))) {dir.create(file.path(image_path, "other", "poster"), recursive = T)}
if(!dir.exists(file.path(image_path, "other", "qc"))) {dir.create(file.path(image_path, "other", "qc"), recursive = T)}
if(!dir.exists(file.path(image_path, "SI"))) {dir.create(file.path(image_path, "SI"), recursive = T)}

# ggplot settings
theme_set(theme_bw())
theme_update(legend.position = "bottom")


##################################################################################################
# COMMON VARIABLES 
##################################################################################################
all_pollutants <- readRDS(file.path(output_data_path, "main_pollutants.rda"))

pollutant_levels <- c("Total", "10-420 nm", "10-100 nm", "20-1,000 nm", "NO2 (ppb)")

version_examples <- c("4", "12",  #visits/site
                      "1", #"3", #1 & 4 seasons
                      "Bus", "Rush", "Bus Adj", "Rush Adj",
                      "H22 L2", "H2 L22" #visit type
                      )
drop_site_types <- "H 8|H 4|H 10|H 14|H 16|H 20|L 8|L 4|L 10|L 14|l 16|L 20"

#for boxplots
bp_min <- 0.025
print(paste("Qmin in tables is", bp_min))
bp_max <- 0.975
print(paste("Qmax in tables is", bp_max))

one_to_one_band <- 0.2

# UW colors
uw_purple <- "#4b2e83"
uw_gold1 <- "#b7a57a" # web: e8e3d3
uw_gold2 <- "#85754d"

ref_model_col <- config$ref_model_col  

##################################################################################################
# LOAD DATA
##################################################################################################
# exposure model crosswalks (cw)
cw <- read.csv(file.path(dt_path, "model_cw.csv")) %>%
  filter(design %in% c("full", "fewer total stops") |
           # these sensitivity & secondary analyses don't have a temporal adjustment
           design == "fewer hours" & grepl("ns10100|no2", model)) %>%
  ## other designs (includes updated temp adj 1)
  bind_rows(read.csv(file.path(dt_path, "other_designs_model_cw_20240516.csv"))) %>%
  select(-model_no) %>%
  filter(variable %in% all_pollutants,
         !grepl(drop_site_types, version))

##################################################################################################
# exposure MODEL EVAL
# exposure model performances for other designs
model_eval_other_designs0 <- readRDS(file.path(dt_path, "other_designs_model_eval_20240516.rda")) %>%
   select(-no_sites)  

# combine  
model_eval <- readRDS(file.path(dt_path, "model_eval.rda")) %>%
  filter(reference == "gs_estimate",
         design %in% c("full", "fewer total stops") |
           # sensitivity/secondary analyses don't have a temporal adjustment
           design == "fewer hours" & grepl("ns10100|no2", model)) %>%
  select(names(model_eval_other_designs0)) %>%
  bind_rows(model_eval_other_designs0) %>%
  left_join(cw, by="model") %>%
  filter(variable %in% all_pollutants) %>%
  #only show a few of the site types
  filter(!grepl(drop_site_types, version)) %>%
  #remove "temp adj 1 vs 2.." since only 1 option here
  mutate(version = gsub("adj 1|adj 2", "adj", version)) %>%
  label_pollutants() %>% label_designs()  %>%
  mutate(
    ref = ifelse((design == "Fewer Visits" & version == "12") |
                   (design == "Fewer Seasons" & version == "4") |
                   (design == "Unbalanced Visits" & version == "H12 L12"), "Balanced/More Data", "Unbalanced/Less Data")) 

#rm(model_eval_other_designs0)

##################################################################################################  
model_covars_reduced <- readRDS(file.path(output_data_path, "model_covars.rda"))

model_covars_extended <- readRDS(file.path(output_data_path, "model_covars_extended.rda"))

# need more complete datset for HEI ("Model 1 - Reduced Model")
cs0 <- readRDS(file.path(output_data_path, "dt_for_cross_sectional_analysis.rda")) %>%
  rename(exposure = avg_0_5_yr) %>%
  right_join(select(cw, model, variable, design, version, campaign), by="model") %>%
  label_pollutants() %>% label_designs() 

cs <- cs0 %>%
  # only keep subjects in extended (main) health models
  drop_na(all_of(model_covars_extended)) 

##################################################################################################
model_coefs0 <- readRDS(file.path(output_data_path, "model_coefs.rda"))  

model_coefs <- model_coefs0 %>% 
  filter(model %in% unique(cs$model)) %>%
  right_join(select(cw, model, variable, design, version, campaign), by="model") %>%
  label_pollutants() %>% label_designs() %>%
   mutate(
    ref = ifelse((design == "Fewer Visits" & version == "12") |
                   (design == "Fewer Seasons" & version == "4") |
                   (design == "Unbalanced Visits" & version == "H12 L12"), "Balanced/More Data", "Unbalanced/Less Data")) 

# sensitivity health analyses
model_coefs_extended <- readRDS(file.path(output_data_path, "model_coefs_extended.rda")) %>% 
  filter(model %in% unique(cs$model)) %>%
  right_join(select(cw, model, variable, design, version, campaign), by="model") %>%
  label_pollutants() %>% label_designs() %>%
   mutate(
    ref = ifelse((design == "Fewer Visits" & version == "12") |
                   (design == "Fewer Seasons" & version == "4") |
                   (design == "Unbalanced Visits" & version == "H12 L12"), "Balanced/More Data", "Unbalanced/Less Data")) 

model_coefs_2010 <- readRDS(file.path(output_data_path, "model_coefs_extended_2010.rda")) %>%
  filter(grepl("all", model),
         term =="avg_0_5_yr") %>%
  left_join(cw, by="model") %>%
  label_pollutants() %>% label_designs()

exclusion_table <- read.csv(file.path(output_data_path, "exclusion_table2024-07-25.csv")) #exclusion_table2024-05-20.csv

#issue 12 reference models
# --> note that this doesn't use the extended covariates (This is for the HEI report where we did not use the extended covariate health models)
model_coefs_issue12 <- readRDS(file.path(output_data_path, "model_coefs_issue12.rda"))

# for age-equivalent calculations
ap_age_equivalents_reduced <- readRDS(file.path(output_data_path, "model_coefs_all.rda")) %>%
  mutate(analysis="Reduced")

ap_age_equivalents <- readRDS(file.path(output_data_path, "model_coefs_extended_all.rda")) %>%
    mutate(analysis="Primary") %>%
  rbind(ap_age_equivalents_reduced) %>%
  filter(grepl("all_", model),
         grepl("avg_0_5_yr|age", term)) %>%
  mutate(term = ifelse(term == "avg_0_5_yr", "ap",
                       ifelse(term == "visit_age_centered75", "age", term)),
         label = paste0(round(est, 3), " (", round(lower, 3), ", ", round(upper, 3), ")")
         ) %>%
   select(analysis, term, est, label, model,
          lower, upper
          ) %>% 
  pivot_wider(names_from = term, values_from = c(est, label,
                                                 lower, upper
                                                 )) %>%
  mutate(
    ap_age_equivalent = est_ap/est_age,
    ap_month_equivalent = ap_age_equivalent*12,
    
    # add CI to age-equivalent AP effect estimate
    ap_age_equivalent_lower = lower_ap/est_age,
    ap_month_equivalent_lower = ap_age_equivalent_lower*12,
    
    ap_age_equivalent_upper = upper_ap/est_age,
    ap_month_equivalent_upper = ap_age_equivalent_upper*12,
    
  ) %>% 
  left_join(cw, by="model") %>%
  label_pollutants() 

##################################################################################################
# annual estimates
# annual estimates: site type
annual_site_type <- readRDS(file.path(dt_path, "annual_site_type.rda")) %>%
  filter(!grepl(drop_site_types, version)) 

# annual estimates: fewer season
annual_fixed_seasons <- readRDS(file.path(dt_path, "fixed_season_annual.rda")) 

# annual estimates: fewer hours & temporal adjustment (for total NS & P-TRAK (no 10-100 nm))
annual_temp_adj <- readRDS(file.path(dt_path, "temporal_adjustment.rda")) %>%
  mutate(visits=12)

# full, fewer visits, fewer hours  
annual_trainint_set0 <- readRDS(file.path(dt_path, "annual_training_set.rda")) 

# annual estimates: fewer hours for 10-100 (no temporal adj available)
annual_non_temp_adj <- annual_trainint_set0 %>%
  filter(design %in% c("fewer hours")) %>% 
  select(location, ns_10_100, no2,
         visits, campaign, design, version, spatial_temporal)
## combine
fewr_hrs <- left_join(annual_temp_adj, annual_non_temp_adj)

# annual estimates: combine everything
annual_trainint_set <- annual_trainint_set0 %>%
  filter(design %in% c("full", "fewer total stops")) %>%
  select(-spatial_temporal) %>%
  bind_rows(annual_site_type) %>%
  bind_rows(annual_fixed_seasons) %>% 
  bind_rows(fewr_hrs)

##################################################################################################
# site predictions
## all data, fewer visits
predictions0 <- readRDS(file.path(dt_path, "UK Predictions", "all_predictions.rda")) %>%
  filter((design %in% c("full", "fewer total stops") |
            design == "fewer hours" & grepl("no2|ns_10_100 ", variable)), #or ns_10_100 
         variable %in% all_pollutants) %>%
  left_join(cw)

# predictions0 %>% filter(location==first(location), variable=="no2", grepl("_01", model), design=="fewer hours")  

# combine predictions
## other designs: unbalanced, temporally adjusted, fixed balanced seasons...
predictions <- readRDS(file.path(dt_path, "UK Predictions", "other_design_predictions_20240516.rda")) %>%
  filter(#grepl("nstot|noscreen", model),
          variable %in% all_pollutants) %>% 
   left_join(cw) %>%
  ## add original designs
  bind_rows(select(predictions0, names(.))) %>%
  filter(model %in% unique(cs$model)) %>%
  
  label_pollutants() %>% label_designs() %>%
  #select(design, version, campaign, location, gs_estimate, campaign_estimate, prediction) %>% 
  mutate(adjusted = ifelse(grepl("adj", model), "Adjusted", "Unadjusted"),
         adjusted = factor(adjusted, levels=c("Unadjusted", "Adjusted")),
         version = gsub("Adj | Adj 1|Adj 2", "", version)) %>%
  mutate(error = prediction - gs_estimate,
         #absolute error
         error = abs(error)) 

# predictions %>% filter(campaign==1, location==first(location)) %>% View()

# number of visits (stops) per site 
stops_w <- readRDS(file.path(dt_path, "stops_used.rda"))

##################################################################################################  
# MEASUREMENT ERROR 
##################################################################################################  
me <- read.csv(file.path(dt_path, "error", "total_bias.csv"))

##################################################################################################
# COMMON VARIABLES 2
##################################################################################################
all_pollutants_labels <- unique(model_eval$variable)
main_pollutants_labels <- all_pollutants_labels[str_detect(all_pollutants_labels, "Total")]
sensitivity_pollutant_labels <- all_pollutants_labels[str_detect(all_pollutants_labels, "Total|10-100|100|20-1,000")]


# x-axis labels for posters
version_levels2 <- setdiff(levels(model_eval$version), c("Business", "Rush")) %>%
  append(c("Business",
         "Business \nTemporally\nAdjusted", "Bus\nTemp\nAdj", "Bus\nAdj",
         "Rush",
         "Rush \nTemporally\nAdjusted", "Rush\nTemp\nAdj", "Rush\nAdj")) %>%
  unique()

design_labels <- unique(model_eval$design) %>% 
  # drop?
  setdiff("All Data")
# arrange designs
design_labels <- levels(model_eval$design)[levels(model_eval$design) %in% design_labels]

local_tz <- "America/Los_Angeles"

```

# Study Dates 


```{r}
range(cs$visitdt)

```

# QC variables

these QC variable values from the exposure (issue 17) & health (issue 12) dataset should be identical or very similar, and they are. 

```{r}
# names(cs)
qc_vars <- c("exp_coverage", "avg_wc_ufp_10_42_MM_05_yr",
           "exact_coverage", "avg_ec_ufp_10_42_MM_05_yr",
           "imp_coverage", "avg_ic_ufp_10_42_MM_05_yr", 
           "imputation_quality", "avg_iq_ufp_10_42_MM_05_yr")

cs %>%
  # person-level dataset
  select(study_id, all_of(qc_vars)) %>%
  distinct() %>%
  pivot_longer(cols = -study_id, names_to = "var", values_to = "value") %>%
  mutate(var=factor(var, levels=qc_vars)) %>%
  group_by(var) %>% 
  # 1 person has 4 QC variables missing
  drop_na(value) %>%
  summarize(
    N = n(),
    Min = min(value),
    Q25 = quantile(value, 0.25),
    Median = median(value),
    Mean = mean(value),
    Q75 = quantile(value, 0.75),
    Max = max(value)
  ) %>%
  kable(caption = "QC variables", digits = 2) %>%
  kable_styling()

# highest IQ value
distinct(cs, study_id, avg_iq_ufp_10_42_MM_05_yr) %>%
  drop_na(avg_iq_ufp_10_42_MM_05_yr) %>%
  summarize(
    n = n(),
    iq0 = sum(avg_iq_ufp_10_42_MM_05_yr==0)/n,
    iq1 = sum(avg_iq_ufp_10_42_MM_05_yr==1)/n,
    iq2 = sum(avg_iq_ufp_10_42_MM_05_yr==2)/n,
  )%>%
  kable(caption = "proportion of participants with varying imputation qualities", digits = 3) %>%
  kable_styling()

```



# Exposure

## visits per site in the "all data" reference campaign
note that these "visits" are based on the NanoScan (different for on-road data based on P-TRAK data)

```{r}
stops_w %>%
  drop_na(ns_total_conc) %>%
  #drop_na(pnc_noscreen) %>%
  
  group_by(location) %>%
  summarize(visits = n()) %>%
  summarize(
    min = min(visits),
    q25 = quantile(visits, 0.25),
    mean = mean(visits),
    q50 = quantile(visits, 0.50),
    max=max(visits)
  ) %>%
  kable(caption = "distribution of visits per stop with NanoScan readings") %>%
  kable_styling()

```



## Site annual averages


```{r}
annual_avgs <- annual_trainint_set %>%
  pivot_longer(cols = starts_with(c("pnc_", "ns_", "no2")), names_to = "variable", values_to = "value") %>%
  #temp adj is only for NS total
  drop_na(value) %>%
  filter(variable %in% all_pollutants) %>%
  label_pollutants(label = "ufp_range") %>%
  label_designs() %>%
  mutate(primary = ifelse(grepl("10-420", variable), "Primary", "Sensitivity"))

annual_avgs %>%
  group_by(design, version, variable) %>%
  summarize(
    n = n(),
    Min = min(value),
    Q25 = quantile(value, 0.25),
    Median = median(value),
    Mean = mean(value),
    Q75 = quantile(value, 0.75),
    Max = max(value)
  ) %>%
  kable(format.args = list(big.mark = ","), digits = 0) %>%
  kable_styling()

print("distribution of annual average site estimates used to develop exposure models")
annual_avgs %>%
  filter(grepl("nm", variable)) %>%
  group_by(version, design, instrument, variable) %>%
  summary_table(var = "value") %>%  
  
  mutate(
    version = gsub("Bus", "Business", version),
             version = gsub(" Adj 1", " \nTemporally\nAdjusted", version),
             version = factor(version, levels= version_levels2)) %>% #View()
  

  ggplot(aes(x=version, col=interaction(instrument, variable, sep = " "))) +
  geom_boxplot(position = "dodge", stat="identity", aes(ymin=Q10, lower=Q25, middle=Q50, upper=Q75, ymax=Q90)) +
  facet_wrap(~design, scales="free_x", switch = "x", ncol = 2)+
  labs(y = "PNC (pt/cm3)", x="Version", col="Measure", fill="Analysis")

ggsave(file.path(image_path, "SI", "site_averages.png"), width = 8, height = 8)

# NO2
annual_avgs %>%
  filter(grepl("no2", variable, ignore.case=T)) %>%
  group_by(version, design, instrument, variable) %>%
  summary_table(var = "value") %>%  
  
  mutate(
    version = gsub("Bus", "Business", version),
             version = gsub(" Adj 1", " \nTemporally\nAdjusted", version),
             version = factor(version, levels= version_levels2)) %>% #View()
  

  ggplot(aes(x=version, #col=interaction(instrument, variable, sep = " ")
             )) +
  geom_boxplot(position = "dodge", stat="identity", aes(ymin=Q10, lower=Q25, middle=Q50, upper=Q75, ymax=Q90)) +
  facet_wrap(~design, scales="free_x", switch = "x", ncol = 2)+
  labs(y = "NO2 (ppb)", x="Version", col="Measure", fill="Analysis")

ggsave(file.path(image_path, "SI", "site_averages_no2.png"), width = 8, height = 8)

```


## Model R2

**do ppl do weighted avgs when calculationg model R2? weighted by amount of ppl near a location (i.e., how representative is a location relative to the cohort of interest?)**

30 campaigns per design-version-pollutant 

```{r}
t2 <- model_eval %>%
  filter(variable %in% main_pollutants_labels) %>%
  group_by(variable, design, version) %>%
  alt_boxplot(var = "MSE_based_R2", min_q=bp_min, max_q=bp_max) %>%
  ungroup()

t2 %>%
  kable(caption = "Distribution of MSE-based R2 for each design", 
        digits = 2) %>%
  kable_styling()

print("dashed line is the All Data design performance")

refs1 <- model_eval %>%
  filter(design=="All Data", 
         grepl("All", design),
         variable %in% all_pollutants_labels #sensitivity_pollutant_labels
         ) %>%
  select(variable, MSE_based_R2, reg_based_R2) %>%
  mutate(variable = ifelse(grepl("Total", variable), "10-420 nm (Primary)", as.character(variable)),
         variable = factor(variable, levels = c("10-420 nm (Primary)", "10-100 nm", "20-1,000 nm", "NO2 (ppb)")))

# label position is close enought for either R2
temp <- data.frame(design=c("Fewer Visits"),
                   label = c("All-Data Reference Model"),
                   #est=c(r2_ref),
                   version = "6") %>%
  mutate(design = factor(design, levels = levels(model_eval$design)))
  




r2_plot <- function(dt=model_eval, y_variable, facet_variable = "10-420|Total", plot_y_label="PNC") {
  ref_r2 <- refs1 %>%
    filter(grepl(facet_variable, variable, ignore.case=T)) %>%
    pull(get(y_variable))
  
  results_table <- dt %>%
    filter(grepl(facet_variable, variable, ignore.case=T), #variable %in% main_pollutants_labels,
           design != "All Data") %>%
      mutate(version = gsub("Bus", "Business", version),
             version = gsub(" Adj", " \nTemporally\nAdjusted", version),
             version = factor(version, levels= version_levels2),
             
             variable = ifelse(grepl("Total", variable), "10-420 nm (Primary)", as.character(variable)),
             variable = factor(variable, levels = c("10-420 nm (Primary)", "10-100 nm", "20-1,000 nm", "NO2 (ppb)")),
             analysis = ifelse(grepl("10-420", variable), "Primary", "Sensitivity")) %>%
    group_by(version, design, ref) %>%
    summary_table(var = y_variable) 
  
  results_plot <- results_table %>%
    ggplot(aes(x=version, fill=ref)) +
    geom_boxplot(stat="identity", aes(ymin=Q10, lower=Q25, middle=Q50, upper=Q75, ymax=Q90)) +
    geom_hline(yintercept = ref_r2, linetype=2, alpha=0.5, ) +
    facet_wrap(~design, scales = "free_x", switch = "x") + 
    geom_label(data=temp, aes(y=ref_r2, label=label), fill= "white",show.legend = F, size=3) +
    
    scale_y_continuous(sec.axis = sec_axis(trans=~./ref_r2, 
                                           name= "% of Reference Model",
                                           labels = scales::percent,
                                           breaks = seq(0,1,.2)), 
                       position="left") +
      
    labs(#y =paste(plot_y_label, "Exposure Model R2")  
          y = bquote(.(plot_y_label) ~ "Exposure Model" ~ R[MSE]^2), 
            
         x = "Design", fill=NULL)

  return(list(plot=results_plot,
              table = results_table)
         )
  }


r2_plot(y_variable = "MSE_based_R2")$plot  
                      
ggsave(file.path(image_path, "r2.png"), width = 8, height = 7)


r2_plot(y_variable = "MSE_based_R2")$table %>%
  select(design, version, Min, Q10, Q25, Q50, Q75, Q90, Max) %>%
  arrange(design) %>%
  mutate_if(is.numeric, ~round(., 2)) %>%
  write.csv(file.path(image_path, "HEI Report", "Appendix", "r2_mse.csv"), row.names = F)


################################################################################
# SI Plots - sensitivity analyses

# Updated: does not have % scale on right side b/c of how its plotted
r2_plot_multifacet <- function(dt = model_eval, y_variable) {
  
  dt %>%
    filter(design != "All Data") %>%
      mutate(version = gsub("Bus", "Business", version),
             version = gsub(" Adj", " \nTemporally\nAdjusted", version),
             version = factor(version, levels= version_levels2),
             
             variable = ifelse(grepl("Total", variable), "10-420 nm (Primary)", as.character(variable)),
             variable = factor(variable, levels = c("10-420 nm (Primary)", "10-100 nm", "20-1,000 nm", "NO2 (ppb)")),
             analysis = ifelse(grepl("10-420", variable), "Primary", "Sensitivity")) %>%
    #filter(variable == facet_variable) %>%
    
    group_by(version, design, variable, ref) %>%
    summary_table(var = y_variable) %>%
    ggplot(aes(x = version, fill = ref)) +
    geom_boxplot(stat = "identity", aes(ymin = Q10, lower = Q25, middle = Q50, upper = Q75, ymax = Q90)) +
    geom_hline(data = refs1, aes(yintercept = MSE_based_R2), linetype = 2, alpha = 0.5) + 
    facet_grid(variable ~ design, scales = "free_x", switch = "both") + 
    # scale_y_continuous(sec.axis = sec_axis(trans = ~./ref_r2, name = NULL,
    #                                        labels = scales::percent,
    #                                        breaks = seq(0, 1, .2)),
    #                    position = "left", limits = c(0, 0.8)) +
    labs(x = NULL,
         fill = NULL)
}

r2_plot_multifacet(y_variable= "MSE_based_R2") + 
  labs(y=expression("Exposure Model" ~ R[MSE]^2))

ggsave(file.path(image_path, "SI", "r2_sensitivity.png"), bg="white", width = 12, height = 12)


# regression-based R2
r2_plot_multifacet(y_variable= "MSE_based_R2") + 
  labs(y=expression("Exposure Model" ~ R[Regression]^2))

ggsave(file.path(image_path, "SI", "r2_reg_based_sensitivity.png"), bg="white", width = 12, height = 12)

```

 


for HEI Report

```{r, eval=F}
# setting this to eval=F since function was updated since this and facet_variable is no longer a parameter

# no NO2
lapply(c("10-420 nm (Primary)", "10-100 nm", "20-1,000 nm"), function(x) {r2_plot_multifacet(y_variable= "MSE_based_R2", facet_variable=x)}) %>%
  ggarrange(plotlist = ., ncol = 1, common.legend = T, legend = "bottom") %>%
  annotate_figure(right = "% of Reference Model",
                  left = text_grob(expression("Exposure Model" ~ R[MSE]^2), rot = 90),
                  bottom = "Design")

ggsave(file.path(image_path, "HEI Report", "Appendix", "r2_sensitivity.png"), bg="white", width = 12, height = 11)

```


table of exposure model R2s relative to the reference R2

* i may have rounded thse to nearest 5% in the paper 

```{r}
model_eval %>%
  filter(grepl(main_pollutants_labels, variable)) %>%  
  mutate(ref_MSE_based_R2 = MSE_based_R2[grepl("_all_", model)],
         proportion_different = (MSE_based_R2-ref_MSE_based_R2)/-ref_MSE_based_R2,
         proportion = MSE_based_R2/ref_MSE_based_R2) %>% 
  group_by(design, version, variable) %>%
  summarize(
    n=n(),
    median_est = median(MSE_based_R2),
    median_percent_of_ref = median(proportion)*100,
    median_percent_difference = median(proportion_different)*100,
  ) %>%
  kable(caption = "summary of each design's R2 as a proportion of the reference and difference", digits = 2) %>%
  kable_styling()

```

 

### temporally adjusted designs

compare 2 min NO2 vs UFP (for temporal adjustmnet)

```{r}
compare_1hr_avgs <- stops_w %>%
  drop_na(ns_total_conc, no2) %>%
  group_by(date, hour) %>%
  summarize(no_2min_visits=n(),
            no2 = mean(no2),
            ns_total_conc = mean(ns_total_conc)
            )

# 1387 hour avg paired readings     #8671 2 min pars
nrow(compare_1hr_avgs)
# R=0.68 
cor(compare_1hr_avgs$ns_total_conc, compare_1hr_avgs$no2, method="pearson")
 
```


scatterplot of paired unadjusted & temporally adjusted campaign R2

```{r}

print(paste0("dotted lines are +-", one_to_one_band*100, "%"))

model_eval %>%
  filter(design == "Fewer Hours",
         variable == main_pollutants_labels) %>%
  select(-c(model, RMSE, reg_based_R2, version_code)) %>% 
  mutate(adjusted = ifelse(grepl("Adj", version, ignore.case=T), "Adjusted", "Unadjusted"),
         version = gsub(" Adj", "", version),
         version = gsub("Bus", "Business", version),
         version = paste(version, "Hours")
         ) %>%   
  pivot_wider(names_from = adjusted, values_from = MSE_based_R2) %>%  
  mutate(performance = ifelse(Adjusted > Unadjusted, "R2 Improved", "R2 Not Improved")) %>%
  
  ggplot(aes(x=Unadjusted, y=Adjusted,)) + 
  facet_wrap(~version, switch="x") +
  geom_abline(slope = c(1), intercept = 0, linetype=2, alpha=0.5) +
  geom_abline(slope = c(1-one_to_one_band, 1+one_to_one_band), intercept = 0, linetype=3, alpha=0.25) +
  geom_smooth(method = "lm") +
  geom_point(aes(col=performance)) +
  #make x y equal size
  theme(aspect.ratio = 1) +
  geom_label(x=Inf, y=Inf, label="1-1", vjust=1, hjust=1, show.legend = F, col="black") +
  labs(shape="Fewer Hours Design", 
       col = "Impact of Temporal Adjustment",
       x = "Exposure Model R2 for\nTemporally Unadjusted Campaign",
       y = "Exposure Model R2 for\nTemporally Adjusted Campaign"
       #x = "Unadjusted Campaign R2",
       #y = "Temporally Adjusted Campaign R2"
       )
  
ggsave(file.path(image_path, "SI", "r2_temp_adj.png"), width = 6, height = 4)

```

compare predictions from unadjusted & adjusted campaigns to observations. does the error increase?

•	E.g. is it lower for unadjusted ~ 75% of the time; •	We want to know for site A, are all/most campaigns made worse 



```{r, eval=T}

predictions %>%
  filter(design == "Fewer Hours") %>%
  group_by(version, adjusted) %>%
  summarize(
    N = n(),
    Min = min(error),
    #Q2.5 = quantile(error, 0.025),
    Q25 = quantile(error, 0.25),
    Q50 = quantile(error, 0.50),
    Q75 = quantile(error, 0.75),
    #Q97.5 = quantile(error, 0.975),
    Max = max(error),
    #IQR = Q75-Q25,
    #range = Max-Min
    ) %>%
  kable(caption = "Distribution of prediction error (prediction - gs_estimate). N=309 sites x 30 campaigns", digits = 0) %>%
  kable_styling()

# arrange sites by concentration & variability
#sample_sites <- sample(unique(other_design_site_predictions$location), size = 50)
# select 10 random sites from each concentration decile (low to high conc sites)
temp <- predictions %>%
  filter(variable == main_pollutants_labels,
         design == "Fewer Hours") %>%
  
  distinct(location, gs_estimate) %>%
  arrange(gs_estimate) %>%
  mutate(concentration_rank = row_number(),
         concentration_rank_group = cut_number(concentration_rank, n=10)) 
# 5 ranodm samples per decile
set.seed(5)
sample_sites <- temp %>%
  group_by(concentration_rank_group) %>%
  slice_sample(n=5) 
  

print(paste("Prediction site errors from", nrow(sample_sites), "random sites. Sites are arranged by site concentration (all-data estimate), with low concentration sites near the bottom."))
#other_design_site_predictions %>%
predictions %>%
  filter(#design == "Fewer Hours",
         grepl("nstot_fewhrs", model)) %>%  
  
  filter(location %in% sample_sites$location) %>%
  mutate(location = factor(location, levels=temp$location),
         
         version = gsub(" Adj", "", version),
         version = gsub("Bus", "Business", version),
         version = paste(version, "Hours")
         ) %>% 
  
  group_by(version, location, adjusted) %>%
  summary_table(var = "error") %>% 
  ggplot(aes(y=location, fill=adjusted)) +
  geom_boxplot(stat="identity", aes(xmin=Q10, xlower=Q25, xmiddle=Q50, xupper=Q75, xmax=Q90)) +
  facet_wrap(~version) +
  geom_vline(xintercept = 0, linetype=2) +
  labs(x="Prediction Error (pt/cm3)",
       fill = "Temporally")
 
ggsave(file.path(image_path, "SI", "temp_adj_site_error_boxplots.png"), width = 7, height = 9)


```





```{r, eval=F}
### poster

# this was for an older version that didn't include e.g. site type or temporal adj
model_eval %>%
  filter(variable %in% main_pollutants_labels,) %>%
  ggplot(aes(x=version, y=MSE_based_R2)) + 
  geom_hline(data=select(filter(t2, design=="All Data"), variable, Q50), aes(yintercept = Q50), linetype=2, alpha=0.5) + 
  geom_boxplot(fill=uw_gold1, col=uw_purple)+
  facet_grid(~design, scales = "free", switch = "both") +
  labs(y = "Exposure Model R2", x = "Design")

ggsave(file.path(image_path, "other", "poster", "exposure r2.png"), width = 4.5, height = 2.5)

```

## Participant Exposures

N_predictions = 65400 participant predictions = 30 campaigns * 2128 participants/campaign

```{r}
t1 <- cs %>%
  filter(variable %in% #main_pollutants_labels
           c(as.character(main_pollutants_labels), c("ns_total_conc"))
           ) %>%
  group_by(variable, design, version) %>%
  
  # --> TEMP B/C MISSING some versions
  drop_na(exposure) %>%
  
  #alt_boxplot(var = "exposure")  
  summarize(
      N = n(),
      Min = min(exposure),
      Q2.5 = quantile(exposure, 0.025),
      Q25 = quantile(exposure, 0.25),
      Q50 = quantile(exposure, 0.50),
      Q75 = quantile(exposure, 0.75),
      Q97.5 = quantile(exposure, 0.975),
      Max = max(exposure),
      IQR = Q75-Q25,
      range = Max-Min
    )

t1 %>%
  kable(caption = "Predicted exposures by design", 
        digits = 0, #format.args = list(big.mark  = ",")
        ) %>%
  kable_styling()

```

```{r}
# alt to above: scatterplots
ref <- cs %>%
  filter(grepl("All Data", design),
         variable %in% all_pollutants_labels #sensitivity_pollutant_labels
         ) %>%
  select(study_id, all_exposure = exposure, all_model = model, variable)

samples <- cs %>%
  filter(!grepl("All Data", design),
         variable %in% all_pollutants_labels, #sensitivity_pollutant_labels,
         #only show some examples
         version %in% version_examples
         ) %>%
  select(study_id, exposure,  model, variable, design, version, campaign)

comp_exp <- left_join(ref, samples)
# x=all_pollutants_labels[1]
lapply(all_pollutants_labels, #sensitivity_pollutant_labels, 
       function(x) {
  
  p <- comp_exp %>%
  filter(variable== x) %>%
    ggplot(aes(x=all_exposure, y=exposure, col=campaign, group=campaign)) +
    facet_wrap_equal(~design+version, ncol=4) + 
    geom_smooth(se = F, size=0.25, alpha=0.4) + 
    geom_smooth(se = F, inherit.aes = F, aes(x=all_exposure, y=exposure), col="black") + 
    geom_abline(slope = 1, intercept = 0, linetype=2, alpha=0.5) +
    geom_abline(slope = 1.25, intercept = 0, linetype=3, alpha=0.3) + 
    geom_abline(slope = 0.75, intercept = 0, linetype=3, alpha=0.3) + 
    #make x y equal size  
    theme(aspect.ratio = 1) +
    
    labs(x = "All Data Exposure Prediction", 
         y= "Sampling Design Exposure Prediction", 
         title = x,
         col="Campaign")

  if(grepl("Total|10-420", x)) {
    p <- p + labs(title = "10-420 nm (Primary)")
    print(p)
    ggsave(file.path(image_path, "SI", paste0("predicted_exposure_",x ,".png")), width = 9, height = 7)
    } else{
      p <- p + labs(title = x)
      print(p)
      ggsave(file.path(image_path, "SI", paste0("predicted_exposure_",x ,".png")), width = 11, height = 7)
    }

})


```

comparison of exposures from the issue 12 model to the all-data HEI model

```{r}
bound0 <- 0.1
print(paste0("NanoScan reference models. dashed line is 1-1 and +-", bound0*100, "%"))
cs %>%
  filter(model == "s_nstot_all_01") %>%
  ggplot(aes(x=exposure, y=cum_exp_ufp_10_42_MM_05_yr)) + 
  geom_point(alpha=0.1) + 
  geom_smooth() +
  geom_abline(slope = 1, intercept = 0, linetype=2, alpha=0.5) +
  geom_abline(slope = c(1+bound0, 1-bound0), intercept = 0, linetype=3, alpha=0.3) + 
  #make x y equal size  
  theme(aspect.ratio = 1) +
  labs(x="s_nstot_all_01", 
       caption = paste0("dashed line is 1-1 and +-", bound0*100, "%"))

ggsave(file.path(image_path, "other", "qc", "all_data_ns_vs_issue12_model.png"), width = 6, height = 6)

print(paste0("P-TRAK reference models. dashed line is 1-1 and +-", bound0*100, "%"))
cs %>%
  filter(model == "s_pncnoscreen_all_01") %>%
  ggplot(aes(x=exposure, y=cum_exp_ufp_20_1k_MM_05_yr)) + 
  geom_point(alpha=0.1) + 
  geom_smooth() +
  geom_abline(slope = 1, intercept = 0, linetype=2, alpha=0.5) +
  geom_abline(slope = c(1+bound0, 1-bound0), intercept = 0, linetype=3, alpha=0.3) + 
  #make x y equal size  
  theme(aspect.ratio = 1) +
  labs(x="s_pncnoscreen_all_01",
       caption = paste0("dashed line is 1-1 and +-", bound0*100, "%"))
ggsave(file.path(image_path, "other", "qc", "all_data_ns_ptrak_issue12_model.png"), width = 6, height = 6)

```


## Prediction Correlations

```{r}
full <- cs %>% 
  filter(grepl("All Data", design),
         variable %in% all_pollutants_labels#sensitivity_pollutant_labels
         ) %>%
  select(study_id, variable, exposure, model) %>%
  rename(full_prediction = exposure)

designs <- filter(cs, 
                  variable %in% all_pollutants_labels, #sensitivity_pollutant_labels  
                  !model %in% unique(full$model),
                  #drop some of these
                  !version %in% c("H4 L20", "H8 L16", "H10 L14", "H14 L10", "H16 L8", "H20 L4")) %>%
  select(study_id, variable, exposure, model, design, version, campaign) %>%
  rename(design_prediction = exposure)

comp <- left_join(select(full, -model), designs)

```

correlations (N= 30 correlations for each design-version)

```{r}
corr_table <- comp %>%
  group_by(model, variable, design, version, campaign) %>%
  summarize(n=n(),
            r = cor(full_prediction, design_prediction))  %>%
  ungroup() %>%
  mutate(
    ref = ifelse((design == "Fewer Visits" & version == "12") |
                   (design == "Fewer Seasons" & version == "4") |
                   (design == "Unbalanced Visits" & version == "H12 L12"), "Balanced/More Data", "Unbalanced/Less Data"))  

corr_table %>%
  group_by(variable, design, version) %>%
  alt_boxplot(var = "r", min_q=bp_min, max_q=bp_max) %>%
  
  kable(caption = "Distribution of Pearson correlations (R) between full campaign and sampling design predictions (N= 30 correlations for each design-version)", 
        digits = 2) %>%
  kable_styling()

# corr_table %>%
#   filter(variable %in% main_pollutants_labels) %>% 
#   ggplot(aes(x=version, y=r, fill=ref)) +
#   geom_boxplot()+
#   facet_grid(~design, scales = "free", switch="both") + 
#   facet_wrap(~design, scales = "free", switch="x") + 
#   labs(y = "Correlation (R)", x = "Design", fill=NULL)
# 
# #ggsave(file.path(image_path, "prediction_correlations.png"), width = 8, height = 6)

#SI
# --> CHECK THAT THIS LOOKS CORRECT IN UPDATE (HAS TEMP ADJ?)
corr_table %>%
  mutate(variable = ifelse(grepl("Total", variable), "10-420 nm (Primary)", as.character(variable)),
           variable = factor(variable, levels = c("10-420 nm (Primary)", "10-100 nm", "20-1,000 nm", "NO2 (ppb)")),
           analysis = ifelse(grepl("10-420", variable), "Primary", "Sensitivity"),
         
         version = gsub("Bus", "Business", version),
         version = gsub(" Adj 1", " \nTemporally\nAdjusted", version),
         version = factor(version, levels= version_levels2)
         ) %>%
  
  group_by(version, design, variable, ref) %>%
  summary_table(var = "r") %>%  
  ggplot(aes(x=version, fill=ref)) +
  geom_boxplot(stat="identity", aes(ymin=Q10, lower=Q25, middle=Q50, upper=Q75, ymax=Q90)) +
  facet_grid(variable~design, scales = "free_x", switch="both") + 
  labs(y = "Correlation (R)", x = "Design", col="Analysis", fill=NULL) #+ scale_x_discrete(guide = guide_axis(n.dodge=2))

ggsave(file.path(image_path, "SI", "prediction_correlations.png"), width = 12, height = 14)

```


```{r}
# HEI Report
corr_table %>%
  filter(!grepl("NO2", variable)) %>%
  mutate(variable = ifelse(grepl("Total", variable), "10-420 nm (Primary)", as.character(variable)),
           variable = factor(variable, levels = c("10-420 nm (Primary)", "10-100 nm", "20-1,000 nm", "NO2 (ppb)")),
           analysis = ifelse(grepl("10-420", variable), "Primary", "Sensitivity"),
         
         version = gsub("Bus", "Business", version),
         version = gsub(" Adj 1", " \nTemporally\nAdjusted", version),
         version = factor(version, levels= version_levels2)
         ) %>%
  
  group_by(version, design, variable, ref) %>%
  summary_table(var = "r") %>%  
  ggplot(aes(x=version, fill=ref)) +
  geom_boxplot(stat="identity", aes(ymin=Q10, lower=Q25, middle=Q50, upper=Q75, ymax=Q90)) +
  facet_grid(variable~design, scales = "free_x", switch="both") + 
  labs(y = "Correlation (R)", x = "Design", col="Analysis", fill=NULL) #+ scale_x_discrete(guide = guide_axis(n.dodge=2))

ggsave(file.path(image_path, "HEI Report", "Appendix", "prediction_correlations.png"), width = 12, height = 14)

```


### poster

```{r, eval=F}
corr_table %>%
  ggplot(aes(x=version, y=r )) +
  geom_boxplot(fill=uw_gold1, col=uw_purple)+
  facet_grid(~design, scales = "free", switch="both") + 
  labs(y = "Correlation (R)",
       x = "Design")

ggsave(file.path(image_path, "other", "poster", "prediction_correlations.png"), width = 12, height = 3)
```


# CASI


```{r}
# person-level dataset
pop <- cs %>%
  filter(variable %in% main_pollutants_labels,
         design== "All Data") %>% 
  group_by(variable) %>%
  mutate(#high_exposure = exposure > median(exposure),
         exposure_group = ifelse(exposure <= quantile(exposure, 0.33), "Low PNC",
                                 ifelse(exposure > quantile(exposure, 0.33) & exposure < quantile(exposure, 0.66), "Medium PNC",
                                        ifelse(exposure >= quantile(exposure, 0.66), "High PNC", NA))),
         exposure_group = factor(exposure_group, levels = c("Low PNC", "Medium PNC", "High PNC")), 
         visit_age = visit_age_centered75+75
         ) %>% 
  distinct(study_id, casi_irt, visit_age, year2, #apoe, 
           male, degree, 
           
           race_white, 
           nses_z_cx,
           
           
           variable, exposure, #high_exposure, 
           exposure_group) #%>% pivot_wider(names_from = variable, values_from = c(exposure_group, exposure))  

names(pop) <- gsub("\\ .*|\\,.*", "", names(pop))

pop %>%
  alt_boxplot(var="casi_irt") %>%
  kable(caption = "Distribution of CASI-IRT", 
        digits = 2) %>%
  kable_styling()

```


# Table 1

## Extended Models

```{r}
# Table1 tutorials: 
## https://cran.r-project.org/web/packages/table1/table1.pdf 
## https://cran.r-project.org/web/packages/table1/vignettes/table1-examples.html
 
# my.render.cont <- function(x) {
#     with(stats.apply.rounding(stats.default(x), digits=0), c("",
#         "Mean (SD)"=sprintf("%s (&plusmn; %s)", MEAN, SD)))
# }
# my.render.cat <- function(x) {
#     c("", sapply(stats.default(x), function(y) with(y,
#         sprintf("%d (%0.0f %%)", FREQ, PCT))))
# }

# table labels
t1_labels <- list(
  labels = list(
    casi_irt = "CASI-IRT",
    visit_age = "Visit Age",
    year2 = "Calendar Year (2 yr Cat)",
    apoe = "APOE e4 Allele",
    male = "Sex",
    degree = "Degree",
    
    race_white = "Race",
    nses_z_cx = "Neighborhood Disadvantage Index (NDI)",
    
    variable = "Pollutant",
    #apoe_available = "Included in Main Analysis",
    #model_wt = "IPW for Modeling Cohort",
    #high_exposure_PNC = "High PNC Exposure  (>Median)",
    exposure_group = "PNC Exposure (pt/cm3)",
    exposure = "Residential PNC (pt/cm3) Exposure"
    ),
  units = list(
    visit_age = "Years"
     ),
  categoricals = list(
    apoe = list('1' = "Carriers",
                '0' = "Non-Carriers"),
    # apoe_available = list('TRUE' = "Included in Main Analysis",
    #                       'FALSE' = "Not Included in Main Analysis"),
    male = list('1' = "Male",
                '0' = "Female"),
     
        degree = list('0' = "None",
                  '1' = "GED/High School",
                  '3' = "Bachelor's",
                  '4' = "Master's",
                  '5' = "Doctorate",
                  '6' = "Other" #, digits.pct=0
                  ),
    race_white = list('1' = "White",
                      '0' = "People of Color"),

    # income_cat = list('1' = "<$35,000",
    #                   '2' = "$35,000 - <$50,000",
    #                   '3' = "$50,000 - <$75,000",
    #                   '4' = ">$75,000"),
    
    high_exposure_PNC = list('FALSE' = "PNC Below Median",
                                   'TRUE' = "PNC Above Median")
    ) 
  ) 

print("Baseline cohort characteristics stratified by whether 5-year PNC exposure at baseline was below or above the median.")
# Main Table 1
t1read(pop, t1_labels#, render.continuous.default = my.render.cont
         ) %>%
  table1(~ visit_age + male + degree + race_white +  nses_z_cx + casi_irt + exposure | exposure_group, data=.)

```

## HEI Report / Reduced Health Models 

Same table as above, but for the reduced model, which is used for the HEI report and has fewer adjustment variables (i.e., less missingness)

```{r}
pop_reduced_health_model <- cs0 %>%
  
  drop_na(all_of(model_covars_reduced)) %>%
  
  filter(variable %in% main_pollutants_labels,
         design== "All Data") %>% 
  group_by(variable) %>%
  mutate(#high_exposure = exposure > median(exposure),
         exposure_group = ifelse(exposure <= quantile(exposure, 0.33), "Low PNC",
                                 ifelse(exposure > quantile(exposure, 0.33) & exposure < quantile(exposure, 0.66), "Medium PNC",
                                        ifelse(exposure >= quantile(exposure, 0.66), "High PNC", NA))),
         exposure_group = factor(exposure_group, levels = c("Low PNC", "Medium PNC", "High PNC")), 
         visit_age = visit_age_centered75+75
         ) %>% 
  distinct(study_id, casi_irt, visit_age, year2, #apoe, 
           male, degree, 
           
           race_white, 
           nses_z_cx,
           
           
           variable, exposure, #high_exposure, 
           exposure_group) #%>% pivot_wider(names_from = variable, values_from = c(exposure_group, exposure))  

names(pop_reduced_health_model) <- gsub("\\ .*|\\,.*", "", names(pop_reduced_health_model))

 # Reduced Health Model
t1read(pop_reduced_health_model, t1_labels#, render.continuous.default = my.render.cont
         ) %>%
  table1(~ visit_age + male + degree + race_white +  nses_z_cx + casi_irt + exposure | exposure_group, data=.)

```



# Health Model Results

```{r}
rectangle_alpha <- 0.15

```


## Beta Estimates 

as a reference, these are the health effect estimates from issue 12. The models are developed slightly different (e.g., have 3 vs 2 PLS components; ~1 extra modeling covariate is available; all available visit data are used in the models rather than conditioning on availability of NS readings)

```{r}
model_coefs_issue12 %>%
  kable(caption = "issue 12 health effect estimates for NS and P-TRAK models", digits=3) %>%
  kable_styling()
```

### paper

The black horizontal line and gray area are the health effect estimate from the full campaign

```{r, fig.height=8, fig.width=8}
get_reference_betas <- function(dt) {
  dt %>%
  filter(design == "All Data",
         variable %in% all_pollutants_labels
         # (variable %in% sensitivity_pollutant_labels | grepl("no2", variable, ignore.case=T))
           ) %>%
  select(variable, est, lower, upper, se, significant) %>%
   mutate(variable2 = variable,
          variable = ifelse(grepl("Total", variable), "10-420 nm", as.character(variable)),
           variable = factor(variable, levels = c("10-420 nm", "10-100 nm", "20-1,000 nm", "NO2 (ppb)")),
           analysis = ifelse(grepl("10-420", variable), "Primary", "Sensitivity"),
          
          ) 
  }

true_betas <- get_reference_betas(model_coefs)
true_betas_extended <- get_reference_betas(model_coefs_extended)
 
true_betas %>%
  select(-variable2) %>%
  mutate_if(is.numeric, ~round(., 3)) %>%
  kable(caption = "reduced model coefficient health estiamtes from the all data campaign") %>%
  kable_styling()

true_betas_extended %>%
  select(-variable2) %>%
  mutate_if(is.numeric, ~round(., 3)) %>%
  kable(caption = "Extended model health coefficient estimates from the all data campaign") %>%
  kable_styling()


print("extended health models")
p_list <- unique(true_betas_extended$variable) 

lapply(p_list, function(x){ 
    
      p <- model_coefs_extended %>%
        mutate(variable = ifelse(grepl("Total", variable, ignore.case=T), "10-420 nm", as.character(variable)),
           variable = factor(variable, levels = c("10-420 nm", "10-100 nm", "20-1,000 nm", "NO2 (ppb)")),
           analysis = ifelse(grepl("10-420", variable), "Primary", "Sensitivity"),
           adjusted = ifelse(grepl("adj", model, ignore.case=T), "Temporally Adjusted", "Unadjusted"),
           adjusted = factor(adjusted, levels = c("Unadjusted", "Temporally Adjusted")),
           version = gsub(" Adj| Adj 1", "", version),
           significant = ifelse(significant==TRUE, "Stat. Sign.", "Not Stat. Sign.")
           ) %>% 
        filter(variable == x,
             design != "All Data",
             version %in% version_examples) %>% 
        
        ggplot(aes(x=campaign, #col=significant, 
                   shape=adjusted)) +
        geom_hline(yintercept = 0, linetype=2, alpha=0.2) +
      geom_rect(inherit.aes = F, data=filter(true_betas_extended, variable==x), aes(ymin = lower, ymax = upper, xmin=-Inf, xmax=Inf), alpha=rectangle_alpha) +
      geom_hline(data=filter(true_betas_extended, variable==x), aes(yintercept = est)) +
      geom_pointrange(aes(y=est, ymin=lower, ymax=upper), size=0.4, 
                      position=position_dodge(width=0.7)) +
      scale_x_continuous(breaks = c(1, seq(10,30,10))) +
      labs(x="Campaign", y = "Health Estimate (95% CI)", shape=NULL, col=NULL)
    
        p <- p + facet_wrap(~design+version, ncol = 2) 
        
        if(!grepl("10-420 nm", x)){p <- p + labs(title=x)}
        
        print(p)
        
        ggsave(file.path(image_path, "SI", paste0("beta_", x, ".png")), width = 12, height = 8)
      })

```

reduced model

```{r}
#p_list <- unique(true_betas_extended$variable) 

lapply(p_list, function(x){ 
    
      p <- model_coefs %>%
        mutate(variable = ifelse(grepl("Total", variable, ignore.case=T), "10-420 nm", as.character(variable)),
           variable = factor(variable, levels = c("10-420 nm", "10-100 nm", "20-1,000 nm", "NO2 (ppb)")),
           analysis = ifelse(grepl("10-420", variable), "Primary", "Sensitivity"),
           adjusted = ifelse(grepl("adj", model, ignore.case=T), "Temporally Adjusted", "Unadjusted"),
           adjusted = factor(adjusted, levels = c("Unadjusted", "Temporally Adjusted")),
           version = gsub(" Adj| Adj 1", "", version),
           significant = ifelse(significant==TRUE, "Stat. Sign.", "Not Stat. Sign.")
           ) %>% 
        filter(variable == x,
             design != "All Data",
             version %in% version_examples) %>% 
        
        ggplot(aes(x=campaign, shape=adjusted)) +
        geom_hline(yintercept = 0, linetype=2, alpha=0.2) +
      geom_rect(inherit.aes = F, data=filter(true_betas, variable==x), aes(ymin = lower, ymax = upper, xmin=-Inf, xmax=Inf), alpha=rectangle_alpha) +
      geom_hline(data=filter(true_betas, variable==x), aes(yintercept = est)) +
      geom_pointrange(aes(y=est, ymin=lower, ymax=upper), size=0.4, 
                      position=position_dodge(width=0.7)) +
      scale_x_continuous(breaks = c(1, seq(10,30,10))) +
      labs(x="Campaign", y = "Health Estimate (95% CI)", shape=NULL, col=NULL)
    
        p <- p + facet_wrap(~design+version, ncol = 2) 
        
        if(!grepl("10-420 nm", x)){p <- p + labs(title=x)}
        
        print(p)
        
        ggsave(file.path(image_path, "SI", paste0("beta_", x, "_reduced_models.png")), width = 12, height = 8)
      })


```





proportion of campaigns showing "significant" results 

```{r}
model_coefs %>%
  #filter(grepl("Total", variable)) %>%
  group_by(design, version, variable) %>%
  summarize(n=n(),
            proportion_of_campaigns_significant = mean(significant)) %>%
  arrange(-proportion_of_campaigns_significant) %>%
  kable(caption = "proportion of 'significant' estimates from campaigns. Main results w/ models not fully adjusted",
        digits = 2) %>%
  kable_styling()

model_coefs_extended %>% 
  group_by(design, version, variable) %>%
    summarize(n=n(),
              proportion_of_campaigns_significant = mean(significant)) %>%
    arrange(-proportion_of_campaigns_significant) %>%
  kable(caption = "extended models - proportion of 'significant' estimates from campaigns",
        digits = 2) %>%
  kable_styling()


```

```{r}
d <- c("All", "Visits", "Seasons", "Hours")

```

## Beta Summary

```{r}
# keep this even though code below is the same to not show facet label for main analysis figure
# true Coefficient estimate
tot_pnc <- filter(true_betas, grepl("Total|10-420", variable)) %>%
  pull(est)
tot_pnc_extended <- filter(true_betas_extended, grepl("Total|10-420", variable)) %>%
  pull(est)

# y_lims <- rbind(model_coefs, model_coefs_extended) %>%# View()
#   filter(grepl("tot_", model)) %>%
#   pull(est) %>%
#   range()

y_lims <- c(config$health_plot$min_y_ci, config$health_plot$max_y_ci)


y_lims_pct <- seq(-2,2,1)


# dt=model_coefs_extended
summarize_betas <- function(dt, grouping_vars = c("design", "version", "variable")) {
  dt %>%
    filter(grepl(main_pollutants_labels, variable)) %>%
    mutate(
      ref_est = est[grepl("_all_", model)],
      proportion_different = (est - ref_est) / ref_est
    ) %>%
    group_by(across(all_of(grouping_vars))) %>%
    summarize(
      n = n(),
      min = min(est),
      Q10 = quantile(est, 0.10),
      Q25 = quantile(est, 0.25),
      median = median(est),
      Q75 = quantile(est, 0.75),
      Q90 = quantile(est, 0.90),
      max = max(est),
      
      IQR = IQR(est),
      
      median_percent_difference = median(proportion_different) * 100,
      .groups = "drop"  # Ensures ungrouped data in the output
    )
}

###############################################################
# grouping_vars = c("design", "version", "variable")
beta_table_reduced <- summarize_betas(model_coefs)  

beta_table_reduced %>%
    select(design, version, n, Min=min, Q10, Q25, Q50=median, Q75, Q90, Max=max) %>%
  kable(caption = "reduced model. summary of each design's beta as a proportion of the reference and difference", 
        digits = 3) %>%
  kable_styling()
  
# for main paper & HEI appendix 
beta_table_reduced %>%
  select(design, version, n, Min=min, Q10, Q25, Q50=median, Q75, Q90, Max=max) %>%
  mutate_if(is.numeric, ~round(.,2)) %>%
  write.csv(., file.path(image_path, "beta_table_reduced_health_model_v1.csv"), row.names = F)


beta_table_extended <- summarize_betas(model_coefs_extended)  

beta_table_extended %>%
    select(design, version, n, median, Q25, Q75) %>%
  kable(caption = "extended model. summary of each design's beta as a proportion of the reference and difference", 
        digits = 3) %>%
  kable_styling()
  
# for main paper & HEI appendix 
beta_table_extended %>%
  select(design, version, n, Min=min, Q10, Q25, Q50=median, Q75, Q90, Max=max) %>%
  mutate_if(is.numeric, ~round(.,2)) %>%
  write.csv(., file.path(image_path, "beta_table_extended_health_model_v1.csv"), row.names = F)
###############################################################
# grouping_vars = c("design", "variable")
beta_table_reduced <- summarize_betas(model_coefs, grouping_vars = c("design", "variable"))  

beta_table_reduced %>%
  select(design, n, median, Q25, Q75) %>%
  kable(caption = "reduced model. summary of each design's beta as a proportion of the reference and difference", 
        digits = 3) %>%
  kable_styling()
  
write.csv(beta_table_reduced, file.path(image_path, "beta_table_reduced_health_model_v2.csv"), row.names = F)


beta_table_extended <- summarize_betas(model_coefs_extended, grouping_vars = c("design", "variable"))  

beta_table_extended %>%
    select(design, n, median, Q25, Q75) %>%
  kable(caption = "extended model. summary of each design's beta as a proportion of the reference and difference", 
        digits = 3) %>%
  kable_styling()
  
write.csv(beta_table_extended, file.path(image_path, "beta_table_extended_health_model_v2.csv"), row.names = F)



####################################################################################
# plots 
# for individual pollutants

beta_summary_plot <- function(dt,
                              true_betas.,
                              variable.="Total|10-420",
                              facet_label="Reduced Health Model",
                              reference_text_offset=0,
                              null_text_offset=0
                              ) {
  true_betas. <- filter(true_betas., grepl(variable., variable)) %>%
    mutate(design="All Data",
           design= factor(design, levels= levels(model_coefs$design)),
           significant = ifelse(significant==TRUE, "Statistically Significant", "Not Statistically Significant"))

  p <- dt %>%
    filter(grepl(variable., variable),
           design != "All Data") %>%
    mutate(version = gsub("Bus", "Business", version),
           version = gsub("Adj 1", "\nTemporally\nAdjusted", version),
           
          #version = gsub(" Adj 1", "\nAdj", version),
          
          version = factor(version, levels=c(version_levels2))
            
          
          
           ) %>% #distinct(version) %>% pull()

    group_by(version, design, variable, ref) %>%
    summary_table(var = "est") %>%
    mutate(facet_label = facet_label) %>%
    ggplot(aes(x=version, fill=ref)) +
    geom_boxplot(stat="identity", aes(ymin=Q10, lower=Q25, middle=Q50, upper=Q75, ymax=Q90)) +
    facet_grid(facet_label~design, scales="free_x", switch="both") +


    annotate("rect", xmin = -Inf, xmax = Inf, ymin = true_betas.$lower, ymax = true_betas.$upper,
           alpha = 0.1, fill = ref_model_col) +
    geom_hline(yintercept = true_betas.$est, linetype=2, col=ref_model_col,) +
    geom_hline(yintercept = 0, linetype=2, col="red") +

    scale_y_continuous(sec.axis = sec_axis(
      #trans=~(.-true_betas.$est)/-true_betas.$est,
      trans=~(.-true_betas.$est)/true_betas.$est,
                                         name= NULL,
                                         labels = scales::percent,
                                         breaks = y_lims_pct),
                     position="left", limits = y_lims) +
        scale_x_discrete(labels = function(x) str_wrap(x, width = 8)) +  
    theme(
      # strip.text = element_text(size = 14),     # Facet labels
      # axis.text.x = element_text(size = 10)     # X-axis labels
      legend.text = element_text(size = 14),      # Legend item text
      legend.title = element_text(size = 14) 
    ) +

    labs(y = NULL,  
         x = NULL, 
       fill= NULL)

  print(p)
  }


reduced_models <- beta_summary_plot(dt=model_coefs, true_betas.=true_betas) + 
  theme(axis.text.x = element_text(angle = 25, hjust = 1, size = 12),
        axis.text.y = element_text(size = 11),
        strip.text = element_text(size = 14))
extended_models <- beta_summary_plot(dt=model_coefs_extended, true_betas.=true_betas_extended, facet_label = "Primary Health Model") +
  theme(axis.text.x = element_text(angle = 25, hjust = 1, size = 12),
        axis.text.y = element_text(size = 11),
        strip.text = element_text(size = 14))

ggarrange(reduced_models,  extended_models, ncol = 1, common.legend = T, legend = "bottom") %>%
  annotate_figure(left = text_grob(expression(beta[m]), rot = 90, size = 18),
                  right = text_grob("% Difference from Reference", rot = -90, size = 18),
                  bottom = "Design")

ggsave(file.path(image_path, "beta_boxplots_main_reduced.png"), bg="white", #width = 12, height = 10
       width=12,#12, 
       height=12#10
       )

  ################################################################################################
# HEI Report results

reduced_models_hei <- beta_summary_plot(dt=model_coefs, true_betas.=true_betas, facet_label = "Model 1") #Main Health Model

extended_models_hei <- beta_summary_plot(dt=model_coefs_extended, true_betas.=true_betas_extended, facet_label = "Model 2") #Extended Health Model

ggarrange(reduced_models_hei,  extended_models_hei, ncol = 1, common.legend = T, legend = "bottom") %>%
  annotate_figure(left = text_grob(expression(beta), rot = 90, size = 16),
                right = text_grob("% Difference from Reference", rot = 270, size = 16))

ggsave(file.path(image_path, "HEI Report", "beta_boxplots_main_reduced.png"),bg = "white", width = 12, height = 10)

################################################################################################
# SI - other pollutants
#y_lims_sensitivity <- range(model_coefs_extended$est)
y_lims_sensitivity <- range(config$health_plot$min_y_ci, config$health_plot$max_y_ci)

y_lims_sensitivity_pct <- seq(-2,2, 2)

print("extended health models")
lapply(c("10-420 nm", "10-100 nm", "20-1,000 nm", "NO2 (ppb)"), 
       function(x) {
  
  true_betas2 <- true_betas_extended %>% 
    filter(variable==x) %>%
    mutate(variable = ifelse(variable == "10-420 nm", "10-420 nm (Primary)", as.character(variable)),
           design="All Data",
           design= factor(design, levels= levels(model_coefs$design)))
 
  # for second axis
  ref_conc <- true_betas2$est 
 
  p <- model_coefs_extended %>%
    mutate(variable = ifelse(grepl("Total", variable), "10-420 nm", as.character(variable))) %>%
     filter(#grepl(x, variable)
      variable==x,
      design != "All Data"
           ) %>%
    mutate(variable = ifelse(grepl("Total|10-420 nm", variable), "10-420 nm (Primary)", as.character(variable)),
           variable = factor(variable, levels = c("10-420 nm (Primary)", "10-100 nm", "20-1,000 nm", "NO2 (ppb)")))  %>%
   
  
    group_by(version, design, variable, ref) %>%
    summary_table(var = "est") %>%
    ggplot(aes(x=version, fill=ref)) +
    geom_boxplot(stat="identity", aes(ymin=Q10, lower=Q25, middle=Q50, upper=Q75, ymax=Q90)) +
    
    facet_grid(variable~design, scales="free", switch="both", space="free") + 
    # geom_hline(yintercept = ref_conc, linetype=1) +
    # geom_hline(yintercept = 0, linetype=2, alpha=0.2) +
    annotate("rect", xmin = -Inf, xmax = Inf, ymin = true_betas2$lower, ymax = true_betas2$upper,
           alpha = 0.1, fill = ref_model_col) +
    geom_hline(yintercept = true_betas2$est, linetype=2, col=ref_model_col,) +
    geom_hline(yintercept = 0, linetype=2, col="red") +
    
    scale_y_continuous(sec.axis = sec_axis(
      #trans=~(.-ref_conc)/-ref_conc, 
      trans=~(.-ref_conc)/ref_conc, 
      name=NULL, labels = scales::percent, breaks = y_lims_sensitivity_pct
                                           ), position="left", limits = y_lims_sensitivity) +
  labs(y=NULL, x = "Design", fill=NULL)
  
  p
  }) %>%
  ggarrange(plotlist = ., ncol = 1, common.legend = T, legend = "bottom") %>%
  annotate_figure(right = "% Difference from Reference", 
                   
                   left = text_grob(expression(beta[m]), 
                                   rot = 90, size = 14),
                  )
  
ggsave(file.path(image_path, "SI", "beta_boxplots.png"), bg="white", width = 10, height = 12)


```

### presentation examples 

```{r}
# presentation example of results for reduced model

# dt=model_coefs
# true_betas.=true_betas
# variable.="Total|10-420"
# facet_label=""
# reference_text_offset=0
# null_text_offset=0

beta_summary_plot_sample <- function(dt=model_coefs,
                              true_betas.= true_betas,
                              variable.="Total|10-420",
                              facet_label="",
                              reference_text_offset=0,
                              null_text_offset=0
                              ) {
  true_betas. <- filter(true_betas., grepl(variable., variable)) %>%
    mutate(design="All Data",
           design= factor(design, levels= levels(model_coefs$design)),
           significant = ifelse(significant==TRUE, "Statistically Significant", "Not Statistically Significant"))

  p <- dt %>%
    filter(grepl(variable., variable),
           !design %in% c("All Data", "Unbalanced Visits"),
           !grepl("Adj", version)
           ) %>%
    mutate(version = gsub("Bus", "Business", version),
           version = gsub("Adj 1", "\nTemporally\nAdjusted", version),
           version = factor(version, levels=version_levels2)) %>%

    group_by(version, design, variable, ref) %>%
    summary_table(var = "est") %>%
    #mutate(facet_label = facet_label) %>%
    ggplot(aes(x=version, fill=ref)) +
    geom_boxplot(stat="identity", aes(ymin=Q10, lower=Q25, middle=Q50, upper=Q75, ymax=Q90)) +
    facet_grid(~design, scales="free_x", switch="both") +


    annotate("rect", xmin = -Inf, xmax = Inf, ymin = true_betas.$lower, ymax = true_betas.$upper,
           alpha = 0.1, fill = ref_model_col) +
    geom_hline(yintercept = true_betas.$est, linetype=2, col=ref_model_col,) +
    geom_hline(yintercept = 0, linetype=2, col="red") +

    scale_y_continuous(sec.axis = sec_axis(
      #trans=~(.-true_betas.$est)/-true_betas.$est,
      trans=~(.-true_betas.$est)/true_betas.$est,
                                         name= NULL,
                                         labels = scales::percent,
                                         breaks = c(0.25, 0, -0.25, -0.5, -0.75, -1)
      ),
                     position="left", 
      #limits = y_lims
      ) +
        scale_x_discrete(labels = function(x) str_wrap(x, width = 8)) +  
    theme(
      axis.title = element_text(size = 16),      # Axis titles
      axis.text = element_text(size = 14),       # Axis text labels
      legend.text = element_text(size = 14),     # Legend text
      legend.title = element_text(size = 14),    # Legend title (if used)
      strip.text = element_text(size = 14)       # Facet label text
    ) +

    labs(y = expression(beta),  
         x = "Restricted Sampling Monitoring Design", 
       fill= NULL)

  print(p)
  }


beta_summary_plot_sample()

ggsave(file.path(image_path, "Other", "presentation", "beta_boxplots_reduced_sample.png"), bg="white", width = 9, height = 5)


```


### presentation with measurement error beta

comparison of corrected & uncorrected beta & CIs along with some example campagins 

```{r}
sample_design_betas <- model_coefs  %>%
    filter(grepl("nstot", model),
           # select a few example versions: 4 visits, 1 season, rh, bh
           grepl("v04|fewhrs_bh_", model)) %>%
    mutate(version = gsub("Bus", "Business", version),
           version = factor(version, levels=version_levels2),
           design = case_when(
             design== "Fewer Visits" ~ "12 Visits",
             design== "Fewer Hours" & version== "Business" ~ "12 Visits (Business Hours)",
             TRUE ~design
           ),
           type="Reduced Monitoring"
           ) %>% 
  
  select(design, version, campaign, est, lower, upper, type)

uncorrected_beta <- true_betas %>%
  filter(variable=="10-420 nm") %>%
  select(est, lower, upper) %>%
  mutate(design="Uncorrected",
         campaign=1,
         type="Measurement Error"
         )

# z-value for a 95% CI 
z_value <- qnorm(0.975)

corrected_beta <- me %>%
  mutate(
    lower = adjusted_beta - z_value*adjusted_SE,
    upper = adjusted_beta + z_value*adjusted_SE
    ) %>%
  select(est = adjusted_beta, lower, upper) %>%
  mutate(design = "Corrected",
         campaign=1,
         type="Measurement Error"
         )
 
example_betas <- bind_rows(sample_design_betas,
                           uncorrected_beta) %>%
  bind_rows(corrected_beta) %>%
  mutate(design = factor(design, levels=c("Corrected",
                                          "Uncorrected",
                                          "12 Visits",
                                          "12 Visits (Business Hours)"))) %>%
  droplevels()

# plot
example_betas %>%
  ggplot(aes(x=campaign, col=design)) +
  facet_grid(cols=vars(type, design), scales = "free_x",drop = T)+
  
  
  geom_hline(yintercept = 0, linetype=2, alpha=0.2) +
  geom_rect(inherit.aes = F, data=select(corrected_beta, -c(design, type)), aes(ymin = lower, ymax = upper, xmin=-Inf, xmax=Inf), alpha=rectangle_alpha, fill=ref_model_col) +
  geom_hline(data=select(corrected_beta, -c(design, type)),
             aes(yintercept = est),
             col=ref_model_col, linetype=2) +
  geom_pointrange(aes(y=est, ymin=lower, ymax=upper), #size=0.4,
                  position=position_dodge(width=0.7), show.legend = F) +
  scale_y_continuous(sec.axis = sec_axis(
    trans=~(.-corrected_beta$est)/corrected_beta$est,
    name= NULL,
    labels = scales::percent,
    # ~6%, 12%, 62% bias
    breaks = c(-1, -0.6, -0.15, #-0.06, 
               0, 0.5)
    )) +
   theme(
    axis.title.x = element_blank(),   # Remove x-axis title
    axis.text.x = element_blank(),    # Remove x-axis text labels
    axis.ticks.x = element_blank(),    # Remove x-axis ticks
    axis.title.y = element_text(size = 14),   # Increase y-axis title size
    axis.text.y = element_text(size = 12),    # Increase y-axis text size
    strip.text = element_text(size = 12)
  ) +
  #scale_color_viridis_d() +
  #scale_color_brewer(palette = "Set1") +
  scale_color_manual(values = c("#009E73", #ref_model_col,
                                "#A6D854",
                                "#E78AC3",
                                "#FC8D62"
                                
                                )) +

  labs(y = expression(beta ~ " (95% CI)"), 
       #x= NULL,#"Campaign", 
       shape=NULL, col=NULL)

ggsave(file.path(image_path, "Other", "presentation", "beta_CI_sample.png"), bg="white", width = 10, height = 4)

```

 


2010+ results with extended adjsuted (primary covariates)

```{r}
model_coefs_2010 %>%
  mutate_if(is.numeric, ~round(., 3)) %>%
  mutate(estimate = paste0(est, " (", lower, ", ", upper, ")")) %>%
  select(variable, estimate) %>%
  arrange(variable) %>%
  kable(caption = "health analyses of 2010+ cohort with extended adjusted covariates",
        digits=3) %>%
  kable_styling()

```

### HEI report 

reduced models are 'primary' models & no NO2
 
```{r}
# SI - other pollutants
y_lims <- range(config$health_plot$min_y_ci, config$health_plot$max_y_ci)
y_lims_sensitivity_pct <- seq(-2,2, 1)

print("extended health models")
# x="20-1,000 nm"
lapply(c("10-420 nm", "10-100 nm", "20-1,000 nm"), 
       function(x) {
  
  true_betas2 <- true_betas %>% 
    filter(variable==x) %>%
    mutate(variable = ifelse(variable == "10-420 nm", "10-420 nm (Primary)", as.character(variable)),
           design="All Data",
           design= factor(design, levels= levels(model_coefs$design)))
 
  # for second axis
  ref_conc <- true_betas2$est 
 
  p <- model_coefs %>%
    mutate(variable = ifelse(grepl("Total", variable), "10-420 nm", as.character(variable))) %>%
     filter(#grepl(x, variable)
      variable==x,
      design != "All Data"
           ) %>%
    mutate(variable = ifelse(grepl("Total|10-420 nm", variable), "10-420 nm (Primary)", as.character(variable)),
           variable = factor(variable, levels = c("10-420 nm (Primary)", "10-100 nm", "20-1,000 nm", "NO2 (ppb)")))  %>%
   
  
    group_by(version, design, variable, ref) %>%
    summary_table(var = "est") %>%
    ggplot(aes(x=version, fill=ref)) +
    geom_boxplot(stat="identity", aes(ymin=Q10, lower=Q25, middle=Q50, upper=Q75, ymax=Q90)) +
    
    facet_grid(variable~design, scales="free", switch="both", space="free") + 
    # geom_hline(yintercept = ref_conc, linetype=1) +
    # geom_hline(yintercept = 0, linetype=2, alpha=0.2) +
    annotate("rect", xmin = -Inf, xmax = Inf, ymin = true_betas2$lower, ymax = true_betas2$upper,
           alpha = 0.1, fill = ref_model_col) +
    geom_hline(yintercept = true_betas2$est, linetype=2, col=ref_model_col,) +
    geom_hline(yintercept = 0, linetype=2, col="red") +
    
    
    scale_y_continuous(sec.axis = sec_axis(trans=~(.-ref_conc)/-ref_conc, name=NULL, labels = scales::percent, 
                                           breaks = y_lims_sensitivity_pct
                                           ), position="left", 
                       limits = y_lims) +
  labs(y=NULL, x = "Design", fill=NULL)
  
  p
  }) %>%
  ggarrange(plotlist = ., ncol = 1, common.legend = T, legend = "bottom") %>%
  annotate_figure(right = "% Different from Reference", 
                  left = "Health Point Estimate")
  
ggsave(file.path(image_path, "HEI Report", "Appendix", "beta_boxplots.png"), bg="white", width = 10, height = 12)




```


### for ppt

```{r, eval=F}

# model_coefs %>%
#   filter(grepl("Total|10-420", variable)) %>%
#   
#   ggplot(aes(x=version, y=est)) +
#   facet_grid(~design, scales="free", switch="both", space="free") + 
#   geom_hline(data=filter(true_betas, grepl("Total", variable)),
#              aes(yintercept = est), linetype=1, alpha=0.5) +
#   geom_hline(yintercept = 0, linetype=2, alpha=0.2) +
#   geom_boxplot(fill=uw_gold1, col=uw_purple)+
#   scale_y_continuous(sec.axis = sec_axis(
#     trans=~(.-tot_pnc)/-tot_pnc, name="% Difference", labels = scales::percent)) +
#   geom_label(data=mutate(filter(true_betas, grepl("Total|10-420", variable)), design="All Data" ) %>% mutate(design= factor(design, levels= levels(model_coefs$design))), 
#                          aes(x=1, y=est, label = round(tot_pnc, 3)), col=uw_purple) +
#   labs(y = "Health Effect Estimate", x = "Design" )


#ggsave(file.path(image_path, "other", "presentation", "beta_summary.png"), width = 6, height = 4)

```

### for poster
```{r, eval=F}
temp <- data.frame(version= 2.5, 
                   label = c("All-Data Reference", "No Association"),
                   est=c(tot_pnc, 0),
                   design = c("Fewer Hours")  
                    
                   ) %>%
  mutate(design = factor(design, levels = levels(model_coefs$design)))

model_coefs %>%
  filter(grepl("Total", variable),
         design != "All Data"
         ) %>%
  group_by(version, design, variable, ref) %>%
  summary_table(var = "est") %>%
  mutate(version = gsub("Bus", "Business", version),
         version = gsub("Adj", "\nTemporally\nAdjusted", version),
         version = factor(version, levels=version_levels2)) %>%
  ggplot(aes(x=version, fill=ref)) +
  geom_boxplot(stat="identity", aes(ymin=Q10, lower=Q25, middle=Q50, upper=Q75, ymax=Q90)) +
  facet_wrap(~design, scales="free_x", switch="x") +
  geom_hline(yintercept = tot_pnc, linetype=1, alpha=0.5, col=uw_purple) +
  geom_hline(yintercept = 0, linetype=2, alpha=0.2) +
  scale_y_continuous(sec.axis = sec_axis(trans=~(.-true_betas.$est)/-true_betas.$est, name="Health Point Estimate (% Different from Reference)", labels = scales::percent), position="right") +

  # scale_y_continuous(sec.axis = sec_axis(trans=~(.-tot_pnc)/-tot_pnc, name="Health Point Estimate\n% Difference", labels = scales::percent), position="right") +
    geom_label(data=temp, aes(y=est, label=label), fill= "white",show.legend = F, size=label_size) +
  geom_label(data=filter(temp, grepl("Reference", label, ignore.case = T)), aes(y=est, label=label), fill= "white",col=uw_purple, show.legend = F, size=label_size) +
  poster_theme +
  scale_fill_manual(values = c(uw_spirit_gold, uw_husky_gold)) +
  labs(y = "Health Point Estimate        ", x = "Design",
       fill= NULL)

ggsave(file.path(image_path, "other", "presentation", "beta_boxplots_main_large_text.png"), width = 15, height = 10)

```



```{r}
bias0 <- model_coefs_extended %>%
  group_by(variable) %>%
  mutate(ref = est[design == "All Data"],
         error = est - ref,
         error_pct = round(error/abs(ref)*100, 1),
         proportion_different = (est-ref)/ref,
         percent_different = proportion_different*100
         ) %>%
  filter(design != "All Data",
         variable %in% main_pollutants_labels) 


bias <- bias0 %>%
  group_by(variable, design, version) %>%
  alt_boxplot(var="percent_different" #var = "error"
              ) 

bias %>% 
  kable(caption = "extended models - Distribution of coefficient difference for each design", 
        #digits = 3
        digits=0
        ) %>%
  kable_styling()

#######################################
# bias0 %>%
#   filter(grepl("Visits|Seasons", design)) %>%
#   alt_boxplot(var = "error") %>%
#   mutate_if(is.numeric, ~round(., 3)) %>% 
#   kable(caption = "Distribution of Coefficient errors for fewer visits & seasons") %>%
#   kable_styling()
# 
# bias0 %>%
#   filter(grepl("Fewer Hours", design)) %>%
#   alt_boxplot(var = "error") %>%
#   mutate_if(is.numeric, ~round(.,3)) %>%
#   kable(caption = "Distribution of Coefficient differences for fewer hours") %>%
#   kable_styling()

```


Percent error in the estimated betas (n=30 campaign models for each design-version)

```{r}

# bias_pct <- bias0 %>%
#   group_by(variable, design, version) %>%
#   alt_boxplot(var = "error_pct") 
# 
# bias_pct %>% 
#   #select(-c(Qmin, Qmax)) %>%
#   mutate_if(is.numeric, ~round(.)) %>%
#   kable(caption = "Distribution of Coefficient difference (%) for each design") %>%
#   kable_styling()


#######################################
# bias0 %>%
#   filter(grepl("Visits|Seasons", design)) %>%
#   alt_boxplot(var = "error_pct") %>%
#   mutate_if(is.numeric, ~round(.)) %>%
#   kable(caption = "Distribution of Coefficient difference (%) for fewer visits & seasons") %>%
#   kable_styling()
# 
# bias0 %>%
#   filter(grepl("Fewer Hours", design)) %>%
#   alt_boxplot(var = "error_pct") %>%
#   mutate_if(is.numeric, ~round(.)) %>%
#   kable(caption = "Distribution of Coefficient errors (%) for fewer hours") %>%
#   kable_styling()

```


## aging equivalents

```{r}

ap_age_equivalents %>%
  select(analysis, variable, label_ap, est_age, ap_month_equivalent, 
         ap_month_equivalent_lower, ap_month_equivalent_upper
         ) %>%
  arrange(variable) %>%
  kable(caption = "Estimated associations between air pollutants (PNC [1,900 pt/cm3], NO2 [3 ppb]) and one year of aging with baseline CASI-IRT score based on all-data air pollution exposure models. Air pollution associations are scaled relative to the age association to contextualize their effects.",
        digits=3, 
        col.names = c("Model", "Pollutant", "Air Pollution Association", "Age Association", "Air Pollution Month Equivalent",
                      "Lower Equivalent", "Upper Equivalent"
                      )
        ) %>%
  kable_styling()


```

# Appendix

## Inclusion/exclusion table

```{r}
# https://cran.r-project.org/web/packages/PRISMAstatement/vignettes/exclusionflowcharts.html 

# --> last row looks weird

print("persons")

flow_exclusions(
  incl_counts = exclusion_table$persons,
  total_label = "Full Cohort",
  incl_labels = exclusion_table$description[2:length(exclusion_table$description)],
  percent_of_total = T,
  format_args = list(big.mark  = ","))

```
