==> R - hei_aim3a/10.1_error_variability.Rmd <==
---
title: "HEI Ch5 - Measurement Error"
author: "Magali Blanco"
date: ' `r Sys.time()` '
output:
  html_document:
    df_print: paged
    toc: yes
    toc_depth: '3'
    number_sections: true
    toc_float: true
    collapsed: false
    smooth_scroll: false
editor_options: 
  chunk_output_type: console
---

```{r echo=F, warning=F, message=F}
######################################################################
# SETUP
######################################################################
knitr::opts_chunk$set(echo = F, 
                      cache=F, cache.comments = F, 
                      message = F, warning = F#, tidy.opts=list(width.cutoff=60), tidy=TRUE 
                      )  

rm(list = ls(all = TRUE))
if (!is.null(sessionInfo()$otherPkgs)) {
  res <- suppressWarnings(
    lapply(paste('package:', names(sessionInfo()$otherPkgs), sep=""),
           detach, character.only=TRUE, unload=TRUE, force=TRUE))}

if (!require("pacman")) install.packages("pacman")
pacman::p_load(tidyverse, lubridate, kableExtra, broom)    

set.seed(1)

dt_path <- file.path("Output", readRDS(file.path("Output", "latest_dt_version.rda")))

output_data_path <- file.path(dt_path, "error")
if(!file.exists(output_data_path)) {dir.create(output_data_path, recursive = T)}

image_path <- file.path("..", "Manuscript", "Images", "v4", "other", "error") 
if(!file.exists(image_path)) {dir.create(image_path, recursive = T)}
# ggplot settings
theme_set(theme_bw())
theme_update(legend.position = "bottom")

######################################################################
# LOAD DATA
######################################################################
main_pollutants <- readRDS(file.path(dt_path, "epi", "20240725", "main_pollutants.rda"))
# --> just look at this here?
main_pollutant1 <- c("ns_total_conc")

# modeling units (1,900 for PNC)
load(file.path(output_data_path, "modeling_units.rdata"))

# merged issue 12 [clean health dataset] & issue 16 (for non-parametric, bootstrapped measurement models)
cs <- readRDS(file.path(dt_path, "epi","20240725", "dt_for_cross_sectional_analysis_error.rda")) %>%
  select(-starts_with("avg_"), -visitdt, avg_0_5_yr) %>%
  #transform units to match modeling units
  mutate_at(vars(avg_0_5_yr), ~./pnc_units)

model_covars <- readRDS(file.path(dt_path, "epi", "20240725","model_covars.rda"))

# reference health models using all the data
ref_models0 <- read.csv(file.path(dt_path, "model_cw.csv")) %>%
  filter(variable %in% main_pollutants,
         design == "full") %>%
  select(variable, model)
ref_models <- ref_models0 %>% pull(model)

# health model with the all data exposure model
# ---> use this model vs what is in issue 12??
## all models with stationary data

models_fp <- file.path(dt_path, "epi", 
                       # --> TEMP?
                       #"20240725",
                       "old/pre20240725",
                       
                       "ref_models.rda")

if(file.exists(models_fp)) {
  models <- readRDS(models_fp)
} else { 
  # note: this file has to be re-run & re-saved if anything changes with the models (e.g., covariates, data)
  models0 <- readRDS(file.path(dt_path, "epi", "20240725","models.rda")) 
  
  ## only keep reference (all data) models
  models <- lapply(models0, function(x) {if(as.character(x$model) %in% ref_models) {x}})  
  rm(models0)
  models <- models[sapply(models,function(x) !is.null(x))]
  names(models) <- ref_models
  saveRDS(models, models_fp) 
}

all_data_coefs <- models$s_nstot_all_01 %>% tidy()  
ref_beta <- all_data_coefs %>%
  filter(term == "avg_0_5_yr") %>%
  pull(estimate)
print(paste("ref beta:", ref_beta))

ref_intercept <- all_data_coefs %>%
  filter(term == "(Intercept)") %>%
  pull(estimate)

ref_se <- all_data_coefs %>%
  filter(term == "avg_0_5_yr") %>%
  pull(std.error)
print(paste("ref SE:", ref_se))

ref_beta_ci <- confint(models$s_nstot_all_01) %>% 
  as.data.frame() %>% rownames_to_column() %>%
  filter(rowname == "avg_0_5_yr")
print("beta CI:")
ref_beta_ci

# monitoring site predictions & true estimates from the full campaign in this study (from 2.3_uk_cv.R)
monitoring_sites <- readRDS(file.path(dt_path, "UK Predictions", "all_predictions.rda")) %>%
  filter(design == "full",
         variable %in% main_pollutants) %>%
  select(location, variable, prediction, gs_estimate) %>%
  #transform units to match modeling units
  mutate_at(vars(prediction, gs_estimate), ~./pnc_units) %>%
  #add model name
  left_join(ref_models0)
unique_monitoring_sites <- unique(monitoring_sites$location)

######################################################################
# COMMON VARIABLES
######################################################################
sim_n <- 500
cohort_n <- length(unique(cs$study_id)) #5409

# Define the confidence level
confidence_level <- 0.95
# Calculate the z-value
z_value <- qnorm((1 + confidence_level) / 2)

######################################################################
# FUNCTIONS
######################################################################
# returns summary table
summary_table <- function(df, value) {
  df <- df %>%
    rename(value = all_of(value)) %>%
    summarize(
    N = n(),
    Min = min(value),
    Q2.5 = quantile(value, 0.025),
    Q25 = quantile(value, 0.25),
    Median = median(value),
    Mean = mean(value),
    Q75 = quantile(value, 0.75),
    Q97.5 = quantile(value, 0.975),
    IQR = IQR(value),
    SD = sd(value),
    Max = max(value))
  
  names(df)[names(df)==value] <- value
  return(df) 
}

```

# Non-parametric 
investigates classical-like measurement error that is associated with uncertainty in the exposure surface

from Keller 2017: "The non-parametric bootstrap resamples monitor locations to reflect variation in the predicted exposure surface derived from different monitor locations and [participants] to capture sampling variability in the epidemiologic analysis arising from different subjects."

```{r}
set.seed(1)
betas_np <- lapply(unique(cs$model), function(x) {
  betas <- tibble(model = x,
                  variable = main_pollutant1, #issue 16 uses total NS
                  beta_type1 = NA,
                  beta_type2 =  NA)
  
  health_model <- as.formula(paste("casi_irt ~ avg_0_5_yr" , "+", paste(model_covars, collapse = "+")))
  
  this_exposure_model <- filter(cs, model == x)
  this_cohort_sample <- sample_n(this_exposure_model, size = cohort_n, replace = T)  
  
  # health effects for the ~5k existing participants
  betas$beta_type1 <- lm(health_model, data = this_exposure_model) %>%
    tidy() %>%
    filter(term=="avg_0_5_yr") %>%
    pull(estimate)
  # health effects for 5k bootstrapped participants
  betas$beta_type2 <- lm(health_model, data = this_cohort_sample) %>%
    tidy() %>%
    filter(term=="avg_0_5_yr") %>%
    pull(estimate)
  
  betas
}) %>%
  bind_rows()

saveRDS(betas_np, file.path(output_data_path, "beta_nonparametric.rda"))

######################################################################
# SUMMARIZE VARIABILITY
######################################################################
betas_np_long <- betas_np %>%
  pivot_longer(starts_with("beta_")) %>%
  mutate(
    bootstrap = "non-parametric",
    description = ifelse(name=="beta_type1", "(a) health inference estimate from different monitor locations and sampling times",
                         ifelse(name=="beta_type2", "(b) health inference estimate from different monitor locations, sampling times, and subjects", NA))) 

betas_np_summary <- betas_np_long %>%
  group_by(bootstrap, description) %>%
  summary_table(value="value")

######################################################################
# RESULTS PLOT
######################################################################
print("Non-parametric analysis comparing the estimated association between PNC (per 1,900 pt/cm3) and CASI-IRT across bootstrapped samples with changing monitoring locations, sampling times, and subjects")
betas_np_long %>%
  mutate(description = ifelse(name=="beta_type1", "(a) monitor locations + sampling times",
                              ifelse(name=="beta_type2", "(b) monitor locations + sampling times + subjects", NA))) %>%
  ggplot(aes(x=description, y=value)) + 
  geom_hline(yintercept = ref_beta, linetype=2) +
  geom_boxplot() +
  labs(y="Beta",
       x = "Added Variability"
  )
ggsave(file.path(image_path, "non_parametric_betas.png"), width = 6, height = 6)

```

# Parametric

looks at the impact of exposure measurement error on the health effect estimate

```{r}
######################################################################
# FULL MODEL COEFFICIENTS & RESIDUALS
######################################################################
# save model coefficients 
# model_coefs <- lapply(models, function(x) {
#   temp <- tidy(x) %>%
#     select(term, estimate) %>%
#     mutate(model = as.character(x$model))}) %>%
#   bind_rows()

# get the SD of each model's residuals
model_residuals <- lapply(models, function(x){
  tibble(residual = x$residuals,
         model = x$model)}) %>%
  bind_rows() %>%
  group_by(model) %>%
  summarize(sd = sd(residual))

######################################################################
# SIMULATE 'CASI' SCORE AT MONITORING LOCATIONS & RUN 'HEALTH' MODELS
######################################################################
# cs_person <- cs %>%
#   select(c(study_id, all_of(model_covars))) %>%
#   distinct()
# 
# # mean/mode model covariate values
# mean_age <- mean(cs_person$visit_age_centered75)
# year_mode <- table(cs_person$year2) %>% as.data.frame() %>%
#   filter(Freq == max(Freq)) %>%
#   pull(Var1)
# mean_sex <-  mean(cs_person$male)
# degree_mode <- table(cs_person$degree) %>% as.data.frame() %>%
#   filter(Freq == max(Freq)) %>%
#   pull(Var1)
# 
# # add average covariate values to monitoring data
# monitoring_sites <- monitoring_sites %>%
#   mutate(visit_age_centered75 = mean_age,
#          year2 = year_mode,
#          male = mean_sex,
#          degree = degree_mode)

# will only be looking at "ns_total_conc" in this analysis
monitoring_sites <- filter(monitoring_sites, variable== main_pollutant1)
this_pollutant <- filter(ref_models0, variable==main_pollutant1)
this_model <-models[[this_pollutant$model]]
residual_sd <- model_residuals$sd[model_residuals$model==this_pollutant$model] 

set.seed(1)
betas_p <- lapply(1:sim_n, function(x){
  betas_bs <- tibble(n = x,
                     variable = main_pollutant1,
                     obs_beta = NA,
                     pred_beta =  NA)
  
  sites_sample_bs <- monitoring_sites %>%
    sample_n(size = cohort_n, replace = T) %>%
    # the resulting beta from lm() later should be similar to the "all data" estimate since casi is simulated using this exposure
    rename(avg_0_5_yr = gs_estimate) %>%
    mutate(
      #predicted CASI
      #simulated_casi = predict(this_model, newdata=.),
      simulated_casi = ref_intercept + ref_beta*avg_0_5_yr,
      #add noise
      simulated_casi = simulated_casi + rnorm(n=cohort_n, mean=0, sd=residual_sd)) %>%
    # rename exposure variable back to its original name
    rename(gs_estimate = avg_0_5_yr)
  
  # estimated beta using site observations/estimates
  ## no need to add other covariates b/c they are identical across the sample. We are interested in differences in the betas
  ## this should produce a similar beta as the "all data" beta estimate
  betas_bs$obs_beta <- lm(simulated_casi ~ gs_estimate, data = sites_sample_bs) %>%
    tidy() %>%
    filter(term=="gs_estimate") %>%
    pull(estimate)
  
  # estimated beta using site predictions (measurement error added by using predictions)
  betas_bs$pred_beta <- lm(simulated_casi ~ prediction, data = sites_sample_bs) %>%
    tidy() %>%
    filter(term=="prediction") %>%
    pull(estimate)
  
  betas_bs 
}) %>%
  bind_rows()

saveRDS(betas_p, file.path(output_data_path, "betas_parametric.rda"))

######################################################################
# RESULTS PLOT
######################################################################
print("estimated association between PNC (per 1,900 pt/cm3) and CASI-IRT, adjusted for age, calendar year, sex, education")
line_bounds <- 0.3
betas_p %>%
  ggplot(aes(x=obs_beta, y=pred_beta)) + 
  geom_abline(slope = 1, intercept = 0, linetype=2, alpha=0.5) + 
  geom_abline(slope = c(1-line_bounds, 1+line_bounds), intercept = 0, linetype=3, alpha=0.3) + 
  geom_point() + 
  geom_smooth() + 
  #make x y equal size
  theme(aspect.ratio = 1) +
  labs(x = "Beta (observed PNC)", y= "Beta (predicted PNC)",
       caption = paste0("1-1 and +-", line_bounds*100, "% line")
  )

ggsave(file.path(image_path, "parametric_beta_obs_pred.png"), width = 6, height = 6)

```

# Bias

```{r}
beta_p_summary_predicted <- betas_p %>%
  mutate(bootstrap = "parametric",
         description = "health inference from predicted PNC at simulated cohort locations (i.e., bootstrapped monitoring locations)") %>%
  group_by(bootstrap, description) %>%
  summary_table(value="pred_beta")


# bias <- betas %>%
#   mutate(bias = pred_beta - obs_beta,
#          bootstrap = "parametric",
#          description = "health inference bias from predicted (vs observed) PNC at monitoring locations") %>%
#   group_by(bootstrap, description) %>%
#   summary_table(value="bias")
#   
# bias
######################################################################
# SUMMARY OF BETA ESTIMATES
######################################################################
betas_summary <- rbind(betas_np_summary, beta_p_summary_predicted) %>% 
  mutate(Ref_Beta = ref_beta,
         Bias = Mean- Ref_Beta) %>%
  select(Bootstrap = bootstrap, Description = description, Mean, SD, Ref_Beta, Bias, everything()) %>%
  ungroup()
  
#betas_summary

total_bias <- betas_summary %>%
  filter(Bootstrap == "parametric" | 
           !grepl("subjects", Description)) %>%
  summarize(Ref_Beta = ref_beta,
            Ref_se = ref_se,
            total_bias = sum(Bias),
            # bias corrected beta
            adjusted_beta = Ref_Beta - total_bias,
            # SE accounting for additional variability from measurement error
            adjusted_SE = filter(betas_summary, grepl("(b)", Description) & 
                                   Bootstrap=="non-parametric") %>%
              pull(SD),
            pct_bias = total_bias/adjusted_beta*100,
            )
  
# total_bias %>%
#   kable(caption = "summary of beta estimates") %>%
#   kable_styling()
#   
# write.csv(total_bias, file.path(output_data_path, "total_bias.csv"), row.names = F)
 
# calculate bias specifically from CL & and BL error
betas_summary <- betas_summary %>%
  mutate(pct_bias = Bias/total_bias$adjusted_beta*100) %>%
  select(Bootstrap, Description, Mean, SD, Ref_Beta, Bias, pct_bias, everything())
  
write.csv(betas_summary, file.path(output_data_path, "beta_summary.csv"), row.names = F)

betas_summary %>%
  kable(caption = "Summary of beta estimates") %>%
  kable_styling()


######################################################################
## compare ref to adjusted SE
adjusted_se <- filter(betas_summary, grepl("(b)", Description) & Bootstrap=="non-parametric") %>% 
  pull(SD)

total_bias <- total_bias %>%
  mutate(
    total_SE_diff = adjusted_se-ref_se,
    pct_SE_diff = total_SE_diff/adjusted_se, # 12.67%
    adjusted_beta_lower = adjusted_beta - z_value * adjusted_se,
    adjusted_beta_upper = adjusted_beta + z_value * adjusted_se

    ) %>%
  select(contains("Ref"), contains("total"), everything())
  
total_bias %>%
  kable(caption = "summary of beta estimates. percent SE is the difference betweed adjusted and unadjsuted SE") %>%
  kable_styling()
  
write.csv(total_bias, file.path(output_data_path, "total_bias.csv"), row.names = F)
 
# print("percent SE difference betweed adjusted and unadjsuted SE")
# (adjusted_se-ref_se)/adjusted_se # 12.67%

```
